Remove Imgui phase 1
