+++ ports/mac/Makefile
@@ -1,8 +1,8 @@
 # Makefile for SDL Reader on macOS (Top-Level - No Explicit Library Build)
-# 
+#
 # This Makefile automatically handles dependencies:
 # - Checks for required system libraries (SDL2, libarchive, webp, etc.)
-# - Downloads and sets up ImGui and MuPDF if needed
+# - Downloads and sets up Nuklear and MuPDF if needed
 # - Provides targets for dependency management
 #
 # Usage:
@@ -31,18 +31,14 @@ TARGET = $(BIN_DIR)/$(TARGET_NAME)
 SRC_FILES = $(filter-out src/power_handler.cpp, $(wildcard $(SRC_DIR)/*.cpp))
 SRCS = $(SRC_FILES) $(wildcard $(CLI_DIR)/*.cpp)
 
-# Dear ImGui source files
-IMGUI_DIR = imgui
-IMGUI_SRCS = $(IMGUI_DIR)/imgui.cpp \
-             $(IMGUI_DIR)/imgui_demo.cpp \
-             $(IMGUI_DIR)/imgui_draw.cpp \
-             $(IMGUI_DIR)/imgui_tables.cpp \
-             $(IMGUI_DIR)/imgui_widgets.cpp \
-             $(IMGUI_DIR)/backends/imgui_impl_sdl2.cpp \
-             $(IMGUI_DIR)/backends/imgui_impl_sdlrenderer2.cpp
+# Nuklear directory (using tg5040's nuklear port)
+NUKLEAR_DIR = ../tg5040/nuklear
 
-# Add ImGui sources to compilation
-SRCS += $(IMGUI_SRCS)
+# Nuklear SDL renderer implementation
+NUKLEAR_IMPL = $(NUKLEAR_DIR)/demo/sdl_renderer/nuklear_sdl_renderer.c
+
+# Add Nuklear implementation to compilation
+SRCS += $(NUKLEAR_IMPL)
 
 # Local MuPDF Build Paths (using our custom build with libarchive support)
 MUPDF_BUILD_PATH = mupdf/build/release
@@ -53,9 +49,9 @@ MUPDF_BUILD_PATH = mupdf/build/release
 # -g: Include debugging information
 # -Imupdf/include: Include path for our custom MuPDF headers with libarchive support
 # -I$(SRC_DIR): Include path for project's own header files (e.g., app.h, document.h)
-# -I$(IMGUI_DIR): Include path for Dear ImGui headers
-# -I$(IMGUI_DIR)/backends: Include path for Dear ImGui backend headers
-CXXFLAGS = -std=c++17 -Wall -Wextra -g -Imupdf/include -I$(SRC_DIR) -I../../include -I$(IMGUI_DIR) -I$(IMGUI_DIR)/backends
+# -I$(NUKLEAR_DIR): Include path for Nuklear headers
+# -I$(NUKLEAR_DIR)/demo/sdl_renderer: Include path for Nuklear SDL renderer backend
+CXXFLAGS = -std=c++17 -Wall -Wextra -g -Imupdf/include -I$(SRC_DIR) -I../../include -I$(NUKLEAR_DIR) -I$(NUKLEAR_DIR)/demo/sdl_renderer
 
 # Include paths for SDL2, SDL2_ttf, and SDL2_image
 # pkg-config --cflags SDL2_ttf SDL2_image added for explicit header paths
@@ -78,18 +74,12 @@ LIBS = -L$(MUPDF_BUILD_PATH) -lmupdf -lmupdf-third \
 # Transforms src/file.cpp into build/file.o and cli/file.cpp into build/file.o
 OBJS = $(patsubst $(SRC_DIR)/%.cpp,$(BUILD_DIR)/%.o,$(filter $(SRC_DIR)/%,$(SRCS))) \
        $(patsubst $(CLI_DIR)/%.cpp,$(BUILD_DIR)/%.o,$(filter $(CLI_DIR)/%,$(SRCS))) \
-       $(BUILD_DIR)/imgui.o \
-       $(BUILD_DIR)/imgui_demo.o \
-       $(BUILD_DIR)/imgui_draw.o \
-       $(BUILD_DIR)/imgui_tables.o \
-       $(BUILD_DIR)/imgui_widgets.o \
-       $(BUILD_DIR)/imgui_impl_sdl2.o \
-       $(BUILD_DIR)/imgui_impl_sdlrenderer2.o
+       $(BUILD_DIR)/nuklear_sdl_renderer.o
 
-.PHONY: all clean build-mupdf setup-mupdf clean-mupdf setup-imgui check-deps install-deps clean-imgui clean-all help
+.PHONY: all clean build-mupdf setup-mupdf clean-mupdf check-deps install-deps clean-all help
 
 # Default target
-all: check-deps $(BIN_DIR) $(BUILD_DIR) setup-imgui setup-mupdf build-mupdf $(TARGET)
+all: check-deps $(BIN_DIR) $(BUILD_DIR) setup-mupdf build-mupdf $(TARGET)
 
 # Check if required system dependencies are installed
 check-deps:
@@ -109,21 +99,6 @@ install-deps:
 	@command -v brew >/dev/null 2>&1 || { echo "Homebrew not found. Please install Homebrew first: https://brew.sh"; exit 1; }
 	brew install pkg-config sdl2 sdl2_ttf sdl2_image libarchive webp
 	@echo "System dependencies installed successfully!"
-
-# Setup Dear ImGui if it doesn't exist or is incomplete
-setup-imgui:
-	@if [ ! -d "imgui" ] || [ ! -f "imgui/imgui.h" ]; then \
-		echo "Setting up Dear ImGui repository..."; \
-		rm -rf imgui; \
-		git clone --depth 1 --branch v1.90.9 https://github.com/ocornut/imgui.git; \
-		if [ ! -f "imgui/imgui.h" ]; then \
-			echo "Error: ImGui setup failed - imgui.h not found"; \
-			exit 1; \
-		fi; \
-		echo "Dear ImGui setup completed successfully"; \
-	else \
-		echo "Dear ImGui repository already exists and is complete"; \
-	fi
 
 # Setup MuPDF repository if it doesn't exist
 setup-mupdf:
@@ -188,27 +163,9 @@ $(BUILD_DIR)/%.o: $(SRC_DIR)/%.cpp | $(BUILD_DIR)
 $(BUILD_DIR)/%.o: $(CLI_DIR)/%.cpp | $(BUILD_DIR)
 	$(CXX) $(CXXFLAGS) $(INC_PATHS) -c $< -o $@
 
-# Compile ImGui source files
-$(BUILD_DIR)/imgui.o: $(IMGUI_DIR)/imgui.cpp | $(BUILD_DIR)
-	$(CXX) $(CXXFLAGS) $(INC_PATHS) -c $< -o $@
-
-$(BUILD_DIR)/imgui_demo.o: $(IMGUI_DIR)/imgui_demo.cpp | $(BUILD_DIR)
-	$(CXX) $(CXXFLAGS) $(INC_PATHS) -c $< -o $@
-
-$(BUILD_DIR)/imgui_draw.o: $(IMGUI_DIR)/imgui_draw.cpp | $(BUILD_DIR)
-	$(CXX) $(CXXFLAGS) $(INC_PATHS) -c $< -o $@
-
-$(BUILD_DIR)/imgui_tables.o: $(IMGUI_DIR)/imgui_tables.cpp | $(BUILD_DIR)
-	$(CXX) $(CXXFLAGS) $(INC_PATHS) -c $< -o $@
-
-$(BUILD_DIR)/imgui_widgets.o: $(IMGUI_DIR)/imgui_widgets.cpp | $(BUILD_DIR)
-	$(CXX) $(CXXFLAGS) $(INC_PATHS) -c $< -o $@
-
-$(BUILD_DIR)/imgui_impl_sdl2.o: $(IMGUI_DIR)/backends/imgui_impl_sdl2.cpp | $(BUILD_DIR)
-	$(CXX) $(CXXFLAGS) $(INC_PATHS) -c $< -o $@
-
-$(BUILD_DIR)/imgui_impl_sdlrenderer2.o: $(IMGUI_DIR)/backends/imgui_impl_sdlrenderer2.cpp | $(BUILD_DIR)
-	$(CXX) $(CXXFLAGS) $(INC_PATHS) -c $< -o $@
+# Compile Nuklear SDL renderer (C file)
+$(BUILD_DIR)/nuklear_sdl_renderer.o: $(NUKLEAR_IMPL) | $(BUILD_DIR)
+	$(CC) $(CXXFLAGS) $(INC_PATHS) -c $< -o $@
 
 # Clean up compiled files and directories
 clean:
@@ -216,18 +173,13 @@ clean:
 	rm -f $(TARGET)
 	rm -f webp-upstream-697749.patch
 
-# Clean ImGui (use with caution - will require re-download)
-clean-imgui:
-	@echo "Removing ImGui directory (will be re-downloaded on next build)..."
-	rm -rf imgui
-
 # Clean MuPDF (use with caution - will require re-download and rebuild)
 clean-mupdf:
 	@echo "Removing MuPDF directory (will be re-downloaded on next build)..."
 	rm -rf mupdf
 
 # Clean everything including downloaded dependencies
-clean-all: clean clean-imgui clean-mupdf
+clean-all: clean clean-mupdf
 	@echo "All build artifacts and dependencies cleaned"
 
 # Show help information
@@ -238,11 +190,9 @@ help:
 	@echo "  all          - Build the complete project (default)"
 	@echo "  check-deps   - Check if system dependencies are installed"
 	@echo "  install-deps - Install system dependencies via Homebrew"
-	@echo "  setup-imgui  - Setup Dear ImGui library"
 	@echo "  setup-mupdf  - Setup MuPDF library"
 	@echo "  build-mupdf  - Build MuPDF with libarchive and WebP support"
 	@echo "  clean        - Remove build artifacts"
-	@echo "  clean-imgui  - Remove ImGui directory"
 	@echo "  clean-mupdf  - Remove MuPDF directory"
 	@echo "  clean-all    - Remove all build artifacts and dependencies"
 	@echo "  help         - Show this help message"
