+++ src/gui_manager.cpp
@@ -1,24 +1,167 @@
 #include "gui_manager.h"
 #include <algorithm>
-#include <cstring> // for strcpy
-#include <imgui.h>
-#include <imgui_internal.h> // For navigation wrapping functions
+#include <cmath>
+#include <cstring>
+#include <cstdio>
+#include <cstdlib>
 
-// Platform-specific ImGui backends
-#ifdef TG5040_PLATFORM
-#include <imgui_impl_sdl.h>         // TG5040 uses v1.85 headers
-#include <imgui_impl_sdlrenderer.h> // Compatible with SDL 2.0.9
-#else
-#include <imgui_impl_sdl2.h> // Modern platforms use v1.89+ headers
-#include <imgui_impl_sdlrenderer2.h>
-#endif
+// Define Nuklear implementation
+#define NK_INCLUDE_FIXED_TYPES
+#define NK_INCLUDE_STANDARD_IO
+#define NK_INCLUDE_STANDARD_VARARGS
+#define NK_INCLUDE_DEFAULT_ALLOCATOR
+#define NK_INCLUDE_VERTEX_BUFFER_OUTPUT
+#define NK_INCLUDE_DEFAULT_FONT
+#define NK_INCLUDE_FONT_BAKING
+#define NK_IMPLEMENTATION
+#include "nuklear.h"
 
-#include <iostream>
+#define NK_SDL_RENDERER_IMPLEMENTATION
+// Include the SDL renderer implementation
+#include "demo/sdl_renderer/nuklear_sdl_renderer.h"
 
+static constexpr SDL_GameControllerButton kAcceptButton = SDL_CONTROLLER_BUTTON_B;
+static constexpr SDL_GameControllerButton kCancelButton = SDL_CONTROLLER_BUTTON_A;
+
+namespace
+{
+static nk_bool nk_combo_begin_label_controller(struct nk_context* ctx, const char* selected, struct nk_vec2 size, bool force_open)
+{
+    if (!ctx || !selected || !ctx->current || !ctx->current->layout)
+    {
+        return 0;
+    }
+
+    struct nk_window* win = ctx->current;
+    struct nk_style* style = &ctx->style;
+
+    enum nk_widget_layout_states state;
+    struct nk_rect header;
+    int is_clicked = nk_false;
+    const struct nk_input* in;
+    const struct nk_style_item* background;
+    struct nk_text text;
+
+    state = nk_widget(&header, ctx);
+    if (state == NK_WIDGET_INVALID)
+    {
+        return 0;
+    }
+
+    in = (win->layout->flags & NK_WINDOW_ROM || state == NK_WIDGET_DISABLED || state == NK_WIDGET_ROM) ? nullptr : &ctx->input;
+    if (nk_button_behavior(&ctx->last_widget_state, header, in, NK_BUTTON_DEFAULT))
+    {
+        is_clicked = nk_true;
+    }
+
+    if (force_open)
+    {
+        ctx->last_widget_state |= NK_WIDGET_STATE_ACTIVED;
+        ctx->last_widget_state |= NK_WIDGET_STATE_HOVER;
+        is_clicked = nk_true;
+    }
+
+    if (ctx->last_widget_state & NK_WIDGET_STATE_ACTIVED)
+    {
+        background = &style->combo.active;
+        text.text = style->combo.label_active;
+    }
+    else if (ctx->last_widget_state & NK_WIDGET_STATE_HOVER)
+    {
+        background = &style->combo.hover;
+        text.text = style->combo.label_hover;
+    }
+    else
+    {
+        background = &style->combo.normal;
+        text.text = style->combo.label_normal;
+    }
+
+    text.text = nk_rgb_factor(text.text, style->combo.color_factor);
+
+    switch (background->type)
+    {
+    case NK_STYLE_ITEM_IMAGE:
+        text.background = nk_rgba(0, 0, 0, 0);
+        nk_draw_image(&win->buffer, header, &background->data.image, nk_rgb_factor(nk_white, style->combo.color_factor));
+        break;
+    case NK_STYLE_ITEM_NINE_SLICE:
+        text.background = nk_rgba(0, 0, 0, 0);
+        nk_draw_nine_slice(&win->buffer, header, &background->data.slice, nk_rgb_factor(nk_white, style->combo.color_factor));
+        break;
+    case NK_STYLE_ITEM_COLOR:
+        text.background = background->data.color;
+        nk_fill_rect(&win->buffer, header, style->combo.rounding, nk_rgb_factor(background->data.color, style->combo.color_factor));
+        nk_stroke_rect(&win->buffer, header, style->combo.rounding, style->combo.border, nk_rgb_factor(style->combo.border_color, style->combo.color_factor));
+        break;
+    }
+
+    {
+        struct nk_rect label;
+        struct nk_rect button;
+        struct nk_rect content;
+        int draw_button_symbol;
+
+        enum nk_symbol_type sym;
+        if (ctx->last_widget_state & NK_WIDGET_STATE_HOVER)
+        {
+            sym = style->combo.sym_hover;
+        }
+        else if (is_clicked)
+        {
+            sym = style->combo.sym_active;
+        }
+        else
+        {
+            sym = style->combo.sym_normal;
+        }
+
+        draw_button_symbol = sym != NK_SYMBOL_NONE;
+
+        button.w = header.h - 2 * style->combo.button_padding.y;
+        button.x = (header.x + header.w - header.h) - style->combo.button_padding.x;
+        button.y = header.y + style->combo.button_padding.y;
+        button.h = button.w;
+
+        content.x = button.x + style->combo.button.padding.x;
+        content.y = button.y + style->combo.button.padding.y;
+        content.w = button.w - 2 * style->combo.button.padding.x;
+        content.h = button.h - 2 * style->combo.button.padding.y;
+
+        text.padding = nk_vec2(0, 0);
+        label.x = header.x + style->combo.content_padding.x;
+        label.y = header.y + style->combo.content_padding.y;
+        label.h = header.h - 2 * style->combo.content_padding.y;
+        if (draw_button_symbol)
+        {
+            label.w = button.x - (style->combo.content_padding.x + style->combo.spacing.x) - label.x;
+        }
+        else
+        {
+            label.w = header.w - 2 * style->combo.content_padding.x;
+        }
+
+        nk_widget_text(&win->buffer, label, selected, nk_strlen(selected), &text, NK_TEXT_LEFT, ctx->style.font);
+
+        if (draw_button_symbol)
+        {
+            nk_draw_button_symbol(&win->buffer, &button, &content, ctx->last_widget_state, &ctx->style.combo.button, sym, style->font);
+        }
+    }
+
+    return nk_combo_begin(ctx, win, size, is_clicked, header);
+}
+} // namespace
+
+// GuiManager implementation
 GuiManager::GuiManager()
 {
-    // Initialize page jump input
-    strcpy(m_pageJumpInput, "1"); // Start with page 1
+    // Initialize font size input with default value
+    int result = snprintf(m_fontSizeInput, sizeof(m_fontSizeInput), "%d", m_currentConfig.fontSize);
+    if (result < 0 || result >= (int) sizeof(m_fontSizeInput))
+    {
+        std::cerr << "Warning: Font size input buffer may be truncated" << std::endl;
+    }
 }
 
 GuiManager::~GuiManager()
@@ -33,124 +176,39 @@ bool GuiManager::initialize(SDL_Window* window, SDL_Renderer* renderer)
         return true;
     }
 
-    // Store the renderer for later use
+    m_window = window;
     m_renderer = renderer;
 
-    // Setup Dear ImGui context (or reuse existing one from FileBrowser)
-    bool isNewContext = (ImGui::GetCurrentContext() == nullptr);
-    if (isNewContext)
+    // Initialize Nuklear with proper SDL renderer backend
+    m_ctx = nk_sdl_init(window, renderer);
+    if (!m_ctx)
     {
-        IMGUI_CHECKVERSION();
-        ImGui::CreateContext();
-    }
-
-    ImGuiIO& io = ImGui::GetIO();
-    io.ConfigFlags |= ImGuiConfigFlags_NavEnableKeyboard; // Enable Keyboard Controls
-    io.ConfigFlags |= ImGuiConfigFlags_NavEnableGamepad;  // Enable Gamepad Controls
-
-#ifdef TG5040_PLATFORM
-    // GuiManager uses 1.98x scaling for better menu fit on 640x480 display (1.8x + 10%)
-    // Always reset to default style first to avoid cumulative scaling
-    ImGui::GetStyle() = ImGuiStyle();       // Reset to default
-    ImGui::StyleColorsDark();               // Apply dark colors
-    ImGui::GetStyle().ScaleAllSizes(1.98f); // Scale to 1.98x from base
-    io.FontGlobalScale = 1.98f;
-#else
-    // Setup Dear ImGui style for non-TG5040 platforms
-    ImGui::StyleColorsDark();
-#endif
-
-    // Setup Platform/Renderer backends
-#ifdef TG5040_PLATFORM
-    // TG5040 uses patched v1.85 SDL Renderer backend for framebuffer compatibility
-    if (!ImGui_ImplSDL2_InitForSDLRenderer(window, renderer))
-    {
-        std::cerr << "Failed to initialize ImGui SDL backend" << std::endl;
+        std::cerr << "Failed to initialize Nuklear SDL context" << std::endl;
         return false;
     }
 
-    if (!ImGui_ImplSDLRenderer_Init(renderer))
+    // Setup font atlas - use simplified approach to avoid API compatibility issues
+    struct nk_font_atlas* atlas;
+    nk_sdl_font_stash_begin(&atlas);
+    nk_sdl_font_stash_end();
+
+    // Set up a modern, attractive color scheme
+    setupColorScheme();
+
+    // Load available fonts from options manager
+    auto fonts = m_optionsManager.getAvailableFonts();
+    m_fontNames.clear();
+    for (const auto& font : fonts)
     {
-        std::cerr << "Failed to initialize ImGui SDL Renderer backend" << std::endl;
-        ImGui_ImplSDL2_Shutdown();
-        return false;
-    }
-#else
-    // Modern platforms use SDL Renderer backend
-    if (!ImGui_ImplSDL2_InitForSDLRenderer(window, renderer))
-    {
-        std::cerr << "Failed to initialize ImGui SDL2 backend" << std::endl;
-        return false;
+        m_fontNames.push_back(font.displayName);
     }
 
-    if (!ImGui_ImplSDLRenderer2_Init(renderer))
-    {
-        std::cerr << "Failed to initialize ImGui SDL Renderer backend" << std::endl;
-        ImGui_ImplSDL2_Shutdown();
-        return false;
-    }
-#endif
+    // Find current font index
+    m_selectedFontIndex = findFontIndex(m_currentConfig.fontName);
+    m_fontDropdownHighlightedIndex = m_selectedFontIndex;
 
     m_initialized = true;
-
-    // Reset ImGui IO state to ensure clean slate (important after file browser)
-    {
-        ImGuiIO& ioReset = ImGui::GetIO();
-        ioReset.WantCaptureKeyboard = false;
-        ioReset.WantCaptureMouse = false;
-        ioReset.NavActive = false;
-        ioReset.NavVisible = false;
-    }
-
-    // Initialize font manager and load config
-    m_currentConfig = m_optionsManager.loadConfig();
-    m_tempConfig = m_currentConfig;
-
-    // Get available fonts
-    const auto& fonts = m_optionsManager.getAvailableFonts();
-
-    // Update UI state and ensure we have a valid font selected
-    if (!m_currentConfig.fontName.empty())
-    {
-        m_selectedFontIndex = findFontIndex(m_currentConfig.fontName);
-        // Ensure the index is valid
-        if (m_selectedFontIndex < 0 || m_selectedFontIndex >= (int) fonts.size())
-        {
-            m_selectedFontIndex = 0;
-            // If the saved font wasn't found, update tempConfig with first available font
-            if (!fonts.empty())
-            {
-                m_tempConfig.fontPath = fonts[0].filePath;
-                m_tempConfig.fontName = fonts[0].displayName;
-            }
-        }
-        else
-        {
-            // Font was found, make sure tempConfig has the font path if it's missing
-            if (!fonts.empty() && m_selectedFontIndex >= 0 && m_selectedFontIndex < (int) fonts.size())
-            {
-                const FontInfo& selectedFont = fonts[m_selectedFontIndex];
-                // Update tempConfig if fontPath is empty (important for saved configs)
-                if (m_tempConfig.fontPath.empty())
-                {
-                    m_tempConfig.fontPath = selectedFont.filePath;
-                    m_tempConfig.fontName = selectedFont.displayName;
-                }
-            }
-        }
-    }
-    else
-    {
-        // No saved config, initialize with first available font
-        if (!fonts.empty())
-        {
-            m_selectedFontIndex = 0;
-            m_tempConfig.fontPath = fonts[0].filePath;
-            m_tempConfig.fontName = fonts[0].displayName;
-            m_tempConfig.fontSize = 12; // default size
-        }
-    }
-
+    std::cout << "Nuklear GUI initialized successfully" << std::endl;
     return true;
 }
 
@@ -161,205 +219,106 @@ void GuiManager::cleanup()
         return;
     }
 
-#ifdef TG5040_PLATFORM
-    ImGui_ImplSDLRenderer_Shutdown();
-    ImGui_ImplSDL2_Shutdown();
-#else
-    ImGui_ImplSDLRenderer2_Shutdown();
-    ImGui_ImplSDL2_Shutdown();
-#endif
-
-    // NOTE: We do NOT call ImGui::DestroyContext() here because:
-    // 1. In browse mode, the file browser will need to create a new ImGui context after this
-    // 2. Destroying and recreating contexts can cause issues with SDL/ImGui state
-    // 3. The context will be cleaned up when the program actually exits (at cleanupSDL)
+    nk_sdl_shutdown();
+    m_ctx = nullptr;
 
     m_initialized = false;
 }
 
 bool GuiManager::handleEvent(const SDL_Event& event)
 {
-    if (!m_initialized)
+    if (!m_initialized || !m_ctx)
     {
         return false;
     }
 
-    // Handle number pad input first if it's visible - don't let ImGui process controller events
-    if (m_showNumberPad)
-    {
-        // Let GUIDE button pass through to app (for quit functionality)
-        if (event.type == SDL_CONTROLLERBUTTONDOWN &&
-            event.cbutton.button == SDL_CONTROLLER_BUTTON_GUIDE)
-        {
-            return false; // Don't handle, let it pass through
-        }
-
-#ifdef TG5040_PLATFORM
-        // Handle B button to close number pad on TG5040
-        // NOTE: On TG5040, SDL reports buttons swapped - SDL BUTTON_A = Physical B
-        if (event.type == SDL_CONTROLLERBUTTONDOWN &&
-            event.cbutton.button == SDL_CONTROLLER_BUTTON_A)
-        {
-            std::cout << "[GuiManager/NumberPad] Physical B pressed (SDL reports BUTTON_A) - closing number pad" << std::endl;
-            hideNumberPad();
-            return true;
-        }
-
-        // Handle joystick button 10 to close number pad on TG5040
-        if (event.type == SDL_JOYBUTTONDOWN && event.jbutton.button == 10)
-        {
-            hideNumberPad();
-            return true;
-        }
-#endif
-
-        // Handle controller buttons (except GUIDE and B on TG5040)
-        if (event.type == SDL_CONTROLLERBUTTONDOWN || event.type == SDL_CONTROLLERBUTTONUP)
-        {
-            // Only handle if it's not GUIDE button
-            if (event.cbutton.button != SDL_CONTROLLER_BUTTON_GUIDE)
-            {
-#ifdef TG5040_PLATFORM
-                // Skip SDL BUTTON_A on TG5040 (physical B) as it's handled above for closing
-                if (event.cbutton.button == SDL_CONTROLLER_BUTTON_A)
-                {
-                    return true; // Already handled
-                }
-#endif
-                return handleNumberPadInput(event);
-            }
-        }
-    }
-
-    // Intercept special buttons when font menu is open
-    if (m_showFontMenu)
-    {
-        // Let GUIDE button pass through to app (for quit functionality)
-        if (event.type == SDL_CONTROLLERBUTTONDOWN &&
-            event.cbutton.button == SDL_CONTROLLER_BUTTON_GUIDE)
-        {
-            return false; // Don't handle, let it pass through
-        }
-
-#ifdef TG5040_PLATFORM
-        // Handle joystick button 10 (toggle menu on TG5040) through mapper
-        if (event.type == SDL_JOYBUTTONDOWN && event.jbutton.button == 10)
-        {
-            // Use mapper if available
-            if (m_buttonMapper)
-            {
-                LogicalButton logicalButton = m_buttonMapper->mapJoystickButton(event.jbutton.button);
-                if (logicalButton == LogicalButton::Extra2)
-                {
-                    toggleFontMenu();
-                    if (m_closeCallback)
-                    {
-                        m_closeCallback();
-                    }
-                    return true; // Event handled
-                }
-            }
-            else
-            {
-                // Fallback to direct handling if no mapper
-                toggleFontMenu();
-                if (m_closeCallback)
-                {
-                    m_closeCallback();
-                }
-                return true; // Event handled
-            }
-        }
-
-        // Handle B button on TG5040 in font menu to close it
-        // NOTE: On TG5040, SDL reports buttons swapped:
-        //   - Physical A button → SDL reports as BUTTON_B
-        //   - Physical B button → SDL reports as BUTTON_A
-        // The ImGui backend has been patched to swap A & B so ImGui sees correct mappings.
-        // Here we just need to intercept physical B (SDL BUTTON_A) to close the menu.
-        if (event.type == SDL_CONTROLLERBUTTONDOWN && event.cbutton.button == SDL_CONTROLLER_BUTTON_A)
-        {
-            std::cout << "[GuiManager/FontMenu] Physical B pressed (SDL BUTTON_A) - closing menu" << std::endl;
-            m_showFontMenu = false;
-            // Reset temp config to current config
-            m_tempConfig = m_currentConfig;
-            m_selectedFontIndex = findFontIndex(m_currentConfig.fontName);
-
-            // Trigger redraw to clear menu from screen
-            if (m_closeCallback)
-            {
-                m_closeCallback();
-            }
-            return true; // Event handled
-        }
-#endif
-    }
-
-    // IMPORTANT: Always let ImGui backend process events to maintain proper internal state
-    // This is critical for gamepad navigation to work correctly when menus are opened
-    // ImGui needs to track controller state even when menus are not visible
-    ImGui_ImplSDL2_ProcessEvent(&event);
-
-    // Only report event as "handled" if ImGui actually wants to capture it AND a menu is visible
-    // This prevents ImGui from consuming events when no menu is open
-    ImGuiIO& io = ImGui::GetIO();
-    bool handled = false;
-
-    // Only capture input if a menu is visible
+    // Debug: Print event information when GUI is visible
     if (m_showFontMenu || m_showNumberPad)
     {
-        // Check if ImGui wants to capture keyboard input
-        if (event.type == SDL_KEYDOWN || event.type == SDL_KEYUP || event.type == SDL_TEXTINPUT)
+        if (event.type == SDL_KEYDOWN)
         {
-            handled = io.WantCaptureKeyboard;
+            std::cout << "[DEBUG] Key DOWN: " << SDL_GetKeyName(event.key.keysym.sym)
+                      << " (scancode: " << event.key.keysym.scancode << ")" << std::endl;
         }
-        // Check if ImGui wants to capture mouse input
-        else if (event.type == SDL_MOUSEBUTTONDOWN || event.type == SDL_MOUSEBUTTONUP ||
-                 event.type == SDL_MOUSEMOTION || event.type == SDL_MOUSEWHEEL)
+        else if (event.type == SDL_KEYUP)
         {
-            handled = io.WantCaptureMouse;
+            std::cout << "[DEBUG] Key UP: " << SDL_GetKeyName(event.key.keysym.sym) << std::endl;
         }
-        // For controller/gamepad, always capture if menu is visible
-        else if (event.type == SDL_CONTROLLERBUTTONDOWN || event.type == SDL_CONTROLLERBUTTONUP ||
-                 event.type == SDL_CONTROLLERAXISMOTION)
+        else if (event.type == SDL_MOUSEBUTTONDOWN)
         {
-            handled = true;
+            std::cout << "[DEBUG] Mouse DOWN at (" << event.button.x << ", " << event.button.y << ")" << std::endl;
+        }
+        else if (event.type == SDL_CONTROLLERBUTTONDOWN)
+        {
+            std::cout << "[DEBUG] Controller button DOWN: " << event.cbutton.button << std::endl;
         }
     }
 
-    return handled;
-}
+    // Handle number pad input if visible
+    if (m_showNumberPad && handleNumberPadInput(event))
+    {
+        std::cout << "[DEBUG] Number pad handled event" << std::endl;
+        return true;
+    }
 
-bool GuiManager::isFontMenuVisible() const
-{
-    return m_showFontMenu;
+    // Handle keyboard navigation for GUI (before nuklear gets the events)
+    if (handleKeyboardNavigation(event))
+    {
+        std::cout << "[DEBUG] Keyboard navigation handled event" << std::endl;
+        return true;
+    }
+
+    // Handle controller input for general GUI navigation
+    if (handleControllerInput(event))
+    {
+        std::cout << "[DEBUG] Controller input handled event" << std::endl;
+        return true;
+    }
+
+    // Use the proper SDL event handler for remaining events
+    bool nuklearHandled = nk_sdl_handle_event(const_cast<SDL_Event*>(&event));
+
+    if (m_showFontMenu || m_showNumberPad)
+    {
+        std::cout << "[DEBUG] nuklear handled event: " << (nuklearHandled ? "YES" : "NO") << std::endl;
+    }
+
+    return nuklearHandled;
 }
 
 void GuiManager::newFrame()
 {
-    if (!m_initialized)
+    if (!m_initialized || !m_ctx)
     {
         return;
     }
 
-#ifdef TG5040_PLATFORM
-    ImGui_ImplSDLRenderer_NewFrame();
-    ImGui_ImplSDL2_NewFrame();
-#else
-    ImGui_ImplSDLRenderer2_NewFrame();
-    ImGui_ImplSDL2_NewFrame();
-#endif
-    ImGui::NewFrame();
+    // Begin input processing - handleEvent will add input during event processing
+    nk_input_begin(m_ctx);
+}
+
+void GuiManager::endFrame()
+{
+    if (!m_initialized || !m_ctx)
+    {
+        return;
+    }
+
+    // End input processing after all events have been handled
+    nk_input_end(m_ctx);
 }
 
 void GuiManager::render()
 {
-    if (!m_initialized || !m_renderer)
+    if (!m_initialized || !m_ctx || !m_renderer)
     {
         return;
     }
 
-    // Render our font menu
+    // End input processing before rendering
+    endFrame();
+
+    // Render font menu
     if (m_showFontMenu)
     {
         renderFontMenu();
@@ -371,20 +330,32 @@ void GuiManager::render()
         renderNumberPad();
     }
 
-    // Render ImGui
-    ImGui::Render();
-    ImDrawData* draw_data = ImGui::GetDrawData();
+    // Render Nuklear with proper anti-aliasing
+    nk_sdl_render(NK_ANTI_ALIASING_ON);
 
-    // Only render draw data if there's actually something to show
-    // This prevents stale menu rendering on TG5040
-    if (draw_data && (m_showFontMenu || m_showNumberPad))
+    // Handle mouse grab state
+    nk_sdl_handle_grab();
+}
+
+bool GuiManager::isFontMenuVisible() const
+{
+    return m_showFontMenu;
+}
+
+void GuiManager::toggleFontMenu()
+{
+    m_showFontMenu = !m_showFontMenu;
+    if (m_showFontMenu)
     {
-#ifdef TG5040_PLATFORM
-        ImGui_ImplSDLRenderer_RenderDrawData(draw_data);
-#else
-        ImGui_ImplSDLRenderer2_RenderDrawData(draw_data, m_renderer);
-#endif
+        m_fontDropdownHighlightedIndex = m_selectedFontIndex;
     }
+    else
+    {
+        m_fontDropdownOpen = false;
+        m_fontDropdownSelectRequested = false;
+        m_fontDropdownCancelRequested = false;
+    }
+    std::cout << "Font menu " << (m_showFontMenu ? "opened" : "closed") << std::endl;
 }
 
 void GuiManager::setCurrentFontConfig(const FontConfig& config)
@@ -392,521 +363,747 @@ void GuiManager::setCurrentFontConfig(const FontConfig& config)
     m_currentConfig = config;
     m_tempConfig = config;
 
-    // Update UI state
-    if (!config.fontName.empty())
+    // Update UI elements
+    int result = snprintf(m_fontSizeInput, sizeof(m_fontSizeInput), "%d", config.fontSize);
+    if (result < 0 || result >= (int) sizeof(m_fontSizeInput))
     {
-        m_selectedFontIndex = findFontIndex(config.fontName);
+        std::cerr << "Warning: Font size input buffer may be truncated" << std::endl;
     }
 
-    // Update reading style index
-    auto allStyles = OptionsManager::getAllReadingStyles();
-    m_selectedStyleIndex = 0;
-    for (size_t i = 0; i < allStyles.size(); i++)
+    result = snprintf(m_zoomStepInput, sizeof(m_zoomStepInput), "%d", config.zoomStep);
+    if (result < 0 || result >= (int) sizeof(m_zoomStepInput))
     {
-        if (allStyles[i] == config.readingStyle)
-        {
-            m_selectedStyleIndex = static_cast<int>(i);
-            break;
-        }
+        std::cerr << "Warning: Zoom step input buffer may be truncated" << std::endl;
     }
+
+    m_selectedFontIndex = findFontIndex(config.fontName);
+    m_fontDropdownHighlightedIndex = m_selectedFontIndex;
 }
 
 bool GuiManager::wantsCaptureMouse() const
 {
-    if (!m_initialized)
-    {
-        return false;
-    }
-
-    ImGuiIO& io = ImGui::GetIO();
-    return io.WantCaptureMouse;
+    // Simple implementation - capture mouse when any window is open
+    return m_showFontMenu || m_showNumberPad;
 }
 
 bool GuiManager::wantsCaptureKeyboard() const
 {
-    if (!m_initialized)
-    {
-        return false;
-    }
-
-    ImGuiIO& io = ImGui::GetIO();
-    return io.WantCaptureKeyboard;
+    // Simple implementation - capture keyboard when any window is open
+    return m_showFontMenu || m_showNumberPad;
 }
 
 void GuiManager::renderFontMenu()
 {
-    // Center the window and make it prominent
-    ImGuiIO& io = ImGui::GetIO();
-    ImVec2 center = ImVec2(io.DisplaySize.x * 0.5f, io.DisplaySize.y * 0.5f);
-
-    ImGui::SetNextWindowPos(center, ImGuiCond_Always, ImVec2(0.5f, 0.5f));
-#ifdef TG5040_PLATFORM
-    // TG5040: Use almost full screen size (640x480 display)
-    ImGui::SetNextWindowSize(ImVec2(io.DisplaySize.x * 0.97f, io.DisplaySize.y * 0.97f), ImGuiCond_Always);
-#else
-    // Other platforms: Fixed width, auto height
-    ImGui::SetNextWindowSize(ImVec2(400, 0), ImGuiCond_Always);
-#endif
-
-    if (!ImGui::Begin("Settings", nullptr, ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove))
-    {
-        ImGui::End();
+    if (!m_ctx)
         return;
+
+    // Debug: Print nuklear input state periodically
+    static int debugCounter = 0;
+    if (debugCounter++ % 60 == 0) // Print every 60 frames (about once per second at 60fps)
+    {
+        std::cout << "[DEBUG] Nuklear state - mouse: ("
+                  << m_ctx->input.mouse.pos.x << ", " << m_ctx->input.mouse.pos.y << ")"
+                  << ", mouse down: " << (m_ctx->input.mouse.buttons[0].down ? "YES" : "NO")
+                  << ", keys pressed: " << m_ctx->input.keyboard.text_len << std::endl;
     }
 
-    // Enable navigation wrapping - pressing up on first item goes to last item, and vice versa
-    ImGuiWindow* window = ImGui::GetCurrentWindow();
-    ImGui::NavMoveRequestTryWrapping(window, ImGuiNavMoveFlags_LoopY);
+    int windowWidth, windowHeight;
+    SDL_GetWindowSize(m_window, &windowWidth, &windowHeight);
 
-    // Get available fonts for validation
-    const auto& fonts = m_optionsManager.getAvailableFonts();
+    // Center the window and make it appropriately sized
+    float centerX = windowWidth * 0.5f;
+    float centerY = windowHeight * 0.5f;
+    float windowW = 450.0f;
+    float windowH = 600.0f;
 
-    // Ensure m_tempConfig has valid font data based on current selection
-    // This handles cases where menu is opened and Apply is clicked without changing selection
-    if (m_tempConfig.fontPath.empty() || m_tempConfig.fontName.empty())
+    // Create settings window
+    if (nk_begin(m_ctx, "Settings", nk_rect(centerX - windowW / 2, centerY - windowH / 2, windowW, windowH),
+                 NK_WINDOW_BORDER | NK_WINDOW_TITLE | NK_WINDOW_NO_SCROLLBAR))
     {
-        if (!fonts.empty() && m_selectedFontIndex >= 0 && m_selectedFontIndex < (int) fonts.size())
+        // Set initial focus to enable keyboard navigation
+        static bool initialFocusSet = false;
+        if (!initialFocusSet)
         {
-            const FontInfo& selectedFont = fonts[m_selectedFontIndex];
-            m_tempConfig.fontPath = selectedFont.filePath;
-            m_tempConfig.fontName = selectedFont.displayName;
-        }
-    }
-
-    // === FONT SETTINGS SECTION ===
-    ImGui::Text("Font Settings");
-    ImGui::Separator();
-
-    // Add informational notice about font settings
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.4f, 0.7f, 1.0f, 1.0f)); // Light blue info color
-    ImGui::TextWrapped("Note: Use Document Default for original fonts. Custom fonts apply to EPUB/MOBI only; PDFs and comics use embedded fonts.");
-    ImGui::PopStyleColor();
-    ImGui::Spacing();
-
-    // Font selection dropdown
-    ImGui::Text("Font Family:");
-
-    if (fonts.empty())
-    {
-        ImGui::TextColored(ImVec4(1, 1, 0, 1), "No fonts found in /fonts directory");
-        ImGui::Text("Please add .ttf or .otf files to the fonts folder");
-    }
-    else
-    {
-        // Ensure selected index is valid
-        if (m_selectedFontIndex < 0 || m_selectedFontIndex >= (int) fonts.size())
-        {
-            m_selectedFontIndex = 0;
+            // This helps with initial focus for keyboard/controller navigation
+            initialFocusSet = true;
         }
 
-        // Build font names list directly from fonts (safer than maintaining separate cache)
-        std::vector<const char*> fontNamesPtrs;
-        fontNamesPtrs.reserve(fonts.size());
-        for (const auto& font : fonts)
+        // Store original styles for highlighting focused widgets
+        struct nk_style_button originalButtonStyle = m_ctx->style.button;
+        struct nk_style_combo originalComboStyle = m_ctx->style.combo;
+        struct nk_style_edit originalEditStyle = m_ctx->style.edit;
+        struct nk_style_slider originalSliderStyle = m_ctx->style.slider;
+        struct nk_style_selectable originalSelectableStyle = m_ctx->style.selectable;
+
+        // === KEYBOARD NAVIGATION DEBUG ===
+        nk_layout_row_dynamic(m_ctx, 20, 1);
+        char focusDebug[64];
+        const char* widgetNames[] = {
+            "Font Dropdown", "Font Size Input", "Font Size Slider", "Zoom Step Input",
+            "Zoom Step Slider", "Page Jump Input", "Go Button", "Numpad Button",
+            "Apply Button", "Reset Button", "Close Button"};
+        snprintf(focusDebug, sizeof(focusDebug), "Focus: %s (%d)",
+                 widgetNames[m_mainScreenFocusIndex % WIDGET_COUNT], m_mainScreenFocusIndex);
+        nk_label_colored(m_ctx, focusDebug, NK_TEXT_LEFT, nk_rgb(255, 255, 0));
+
+        // === FONT SETTINGS SECTION ===
+        nk_layout_row_dynamic(m_ctx, 25, 1);
+        nk_label(m_ctx, "Font Settings", NK_TEXT_LEFT);
+
+        // Separator line
+        nk_layout_row_dynamic(m_ctx, 1, 1);
+        struct nk_rect bounds = nk_widget_bounds(m_ctx);
+        struct nk_command_buffer* canvas = nk_window_get_canvas(m_ctx);
+        nk_stroke_line(canvas, bounds.x, bounds.y, bounds.x + bounds.w, bounds.y, 1.0f, nk_rgb(100, 100, 100));
+
+        const auto& fonts = m_optionsManager.getAvailableFonts();
+        if (fonts.empty())
         {
-            fontNamesPtrs.push_back(font.displayName.c_str());
-        }
-
-        // Set focus to font dropdown when menu is first opened
-        if (m_justOpenedFontMenu)
-        {
-            ImGui::SetKeyboardFocusHere();
-            m_justOpenedFontMenu = false; // Reset flag
-        }
-
-        if (ImGui::Combo("##FontFamily", &m_selectedFontIndex, fontNamesPtrs.data(), fontNamesPtrs.size()))
-        {
-            // Font selection changed - ensure index is still valid
-            if (m_selectedFontIndex >= 0 && m_selectedFontIndex < (int) fonts.size())
-            {
-                const FontInfo& selectedFont = fonts[m_selectedFontIndex];
-                m_tempConfig.fontPath = selectedFont.filePath;
-                m_tempConfig.fontName = selectedFont.displayName;
-            }
-            else
-            {
-                // Reset to first font if index became invalid
-                m_selectedFontIndex = 0;
-                if (!fonts.empty())
-                {
-                    m_tempConfig.fontPath = fonts[0].filePath;
-                    m_tempConfig.fontName = fonts[0].displayName;
-                }
-            }
-        }
-    }
-
-    ImGui::Spacing();
-
-    // Font size slider only
-    ImGui::Text("Font Size: %d pt", m_tempConfig.fontSize);
-    int tempSize = m_tempConfig.fontSize;
-    if (ImGui::SliderInt("##FontSizeSlider", &tempSize, 8, 72))
-    {
-        m_tempConfig.fontSize = tempSize;
-        m_fontSizeChanged = true;
-    }
-
-    ImGui::Spacing();
-    ImGui::Spacing();
-
-    // === READING STYLE SECTION ===
-    ImGui::Text("Reading Style");
-    ImGui::Separator();
-
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.4f, 0.7f, 1.0f, 1.0f)); // Light blue info color
-    ImGui::TextWrapped("Choose a color theme for comfortable reading. Applies to EPUB/MOBI only.");
-    ImGui::PopStyleColor();
-    ImGui::Spacing();
-
-    ImGui::Text("Color Theme:");
-
-    // Get all available reading styles
-    auto allStyles = OptionsManager::getAllReadingStyles();
-    std::vector<const char*> styleNames;
-    styleNames.reserve(allStyles.size());
-    for (const auto& style : allStyles)
-    {
-        styleNames.push_back(OptionsManager::getReadingStyleName(style));
-    }
-
-    // Ensure selected index is valid
-    if (m_selectedStyleIndex < 0 || m_selectedStyleIndex >= (int) allStyles.size())
-    {
-        m_selectedStyleIndex = 0;
-    }
-
-    if (ImGui::Combo("##ReadingStyle", &m_selectedStyleIndex, styleNames.data(), styleNames.size()))
-    {
-        // Style selection changed
-        m_tempConfig.readingStyle = allStyles[m_selectedStyleIndex];
-    }
-
-    ImGui::Spacing();
-    ImGui::Spacing();
-
-    // === ZOOM SETTINGS SECTION ===
-    ImGui::Text("Zoom Settings");
-    ImGui::Separator();
-
-    ImGui::Text("Zoom Step: %d%%", m_tempConfig.zoomStep);
-    ImGui::SameLine();
-    ImGui::TextColored(ImVec4(0.7f, 0.7f, 0.7f, 1), "Adjust the amount of zoom per step");
-
-    // Zoom step slider only
-    int tempZoomStep = m_tempConfig.zoomStep;
-    if (ImGui::SliderInt("##ZoomStepSlider", &tempZoomStep, 1, 50))
-    {
-        m_tempConfig.zoomStep = tempZoomStep;
-    }
-
-    ImGui::Spacing();
-    ImGui::Spacing();
-
-    // === PAGE NAVIGATION SECTION ===
-    ImGui::Text("Page Navigation");
-    ImGui::Separator();
-
-    // Current page display
-    ImGui::Text("Current Page: %d / %d", m_currentPage + 1, m_pageCount);
-
-    ImGui::Spacing();
-
-    // Edge progress bar option
-    bool tempDisableEdgeBar = m_tempConfig.disableEdgeProgressBar;
-    if (ImGui::Checkbox("Disable Edge Progress Bar", &tempDisableEdgeBar))
-    {
-        m_tempConfig.disableEdgeProgressBar = tempDisableEdgeBar;
-    }
-
-    ImGui::SameLine();
-    // Make the help icon a small button so it can be focused with D-pad
-    ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0, 0, 0, 0));                    // Transparent background
-    ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4(0.2f, 0.2f, 0.2f, 0.5f)); // Slight highlight on hover
-    ImGui::PushStyleColor(ImGuiCol_ButtonActive, ImVec4(0.2f, 0.2f, 0.2f, 0.5f));  // Slight highlight when active
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.7f, 0.7f, 0.7f, 1));             // Gray text
-    if (ImGui::SmallButton("(?)##EdgeProgressBarHelp") || ImGui::IsItemFocused())
-    {
-        ImGui::SetTooltip("When enabled, panning at page edges will change pages instantly\nwithout delay. When disabled (default), hold at edge for 300ms.");
-    }
-    ImGui::PopStyleColor(4);
-
-    ImGui::Spacing();
-
-    bool tempShowMinimap = m_tempConfig.showDocumentMinimap;
-    if (ImGui::Checkbox("Show Document Minimap", &tempShowMinimap))
-    {
-        m_tempConfig.showDocumentMinimap = tempShowMinimap;
-    }
-    ImGui::SameLine();
-    ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0, 0, 0, 0));
-    ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4(0.2f, 0.2f, 0.2f, 0.5f));
-    ImGui::PushStyleColor(ImGuiCol_ButtonActive, ImVec4(0.2f, 0.2f, 0.2f, 0.5f));
-    ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(0.7f, 0.7f, 0.7f, 1));
-    if (ImGui::SmallButton("(?)##MinimapHelp") || ImGui::IsItemFocused())
-    {
-        ImGui::SetTooltip("Show a miniature page overlay when zoomed in. This helps visualize\nwhich part of the page is currently visible.");
-    }
-    ImGui::PopStyleColor(4);
-
-    ImGui::Spacing();
-
-    ImGui::Text("Jump to Page:");
-    ImGui::SetNextItemWidth(100);
-    if (ImGui::InputText("##PageJump", m_pageJumpInput, sizeof(m_pageJumpInput), ImGuiInputTextFlags_CharsDecimal))
-    {
-        // Input is being edited, but we don't apply until Go button is pressed
-    }
-
-    // Validate page input
-    bool validPageInput = false;
-    int targetPage = std::atoi(m_pageJumpInput);
-    if (targetPage >= 1 && targetPage <= m_pageCount)
-    {
-        validPageInput = true;
-    }
-
-    // Go button on same line as textbox
-    ImGui::SameLine();
-    if (!validPageInput)
-    {
-        ImGui::BeginDisabled();
-    }
-    if (ImGui::Button("Go"))
-    {
-        if (validPageInput && m_pageJumpCallback)
-        {
-            m_pageJumpCallback(targetPage - 1); // Convert to 0-based
-        }
-    }
-    if (!validPageInput)
-    {
-        ImGui::EndDisabled();
-    }
-
-    // Number pad button on same line as textbox
-    ImGui::SameLine();
-    if (ImGui::Button("Number Pad"))
-    {
-        // Show on-screen number pad for controller input
-        showNumberPad();
-    }
-
-    // Show validation message if needed
-    if (!validPageInput && targetPage != 0)
-    {
-        ImGui::TextColored(ImVec4(1, 0.4f, 0.4f, 1), "Invalid page number");
-    }
-
-    ImGui::Spacing();
-    ImGui::Separator();
-    ImGui::Spacing();
-
-    // === BUTTONS SECTION ===
-    bool hasValidFont = !fonts.empty() && m_selectedFontIndex >= 0 && m_selectedFontIndex < (int) fonts.size();
-
-    // Check if font or style settings have changed
-    bool fontSettingsChanged = (m_tempConfig.fontName != m_currentConfig.fontName ||
-                                m_tempConfig.fontSize != m_currentConfig.fontSize ||
-                                m_tempConfig.readingStyle != m_currentConfig.readingStyle);
-
-    // Show warning if font settings changed
-    if (fontSettingsChanged)
-    {
-        ImGui::PushStyleColor(ImGuiCol_Text, ImVec4(1.0f, 0.8f, 0.0f, 1.0f)); // Yellow/orange warning color
-#ifdef TG5040_PLATFORM
-        ImGui::TextWrapped("Warning: Applying font/style changes will reload the document. This may take several seconds on TG5040.");
-#else
-        ImGui::TextWrapped("Note: Applying font/style changes will reload the document.");
-#endif
-        ImGui::PopStyleColor();
-        ImGui::Spacing();
-    }
-
-    if (!hasValidFont)
-    {
-        ImGui::BeginDisabled();
-    }
-
-    if (ImGui::Button("Apply", ImVec2(90, 40)))
-    {
-        std::cout << "Apply button clicked!" << std::endl;
-        if (hasValidFont && m_fontApplyCallback)
-        {
-            // Ensure m_tempConfig has valid font data based on current selection
-            // This is critical if the user opens the menu and clicks Apply without changing the font
-            if (m_tempConfig.fontPath.empty() || m_tempConfig.fontName.empty())
-            {
-                if (m_selectedFontIndex >= 0 && m_selectedFontIndex < (int) fonts.size())
-                {
-                    const FontInfo& selectedFont = fonts[m_selectedFontIndex];
-                    m_tempConfig.fontPath = selectedFont.filePath;
-                    m_tempConfig.fontName = selectedFont.displayName;
-                    std::cout << "DEBUG: Fixed empty font config - set to: " << m_tempConfig.fontName << std::endl;
-                }
-            }
-
-            std::cout << "DEBUG: Applying config - fontName: '" << m_tempConfig.fontName
-                      << "', fontPath: '" << m_tempConfig.fontPath
-                      << "', fontSize: " << m_tempConfig.fontSize
-                      << ", readingStyle: " << static_cast<int>(m_tempConfig.readingStyle) << std::endl;
-
-            // Update current config
-            m_currentConfig = m_tempConfig;
-
-            // Note: Config will be saved after successful application in app.cpp
-            // Don't save here to avoid double-saving and saving before changes are applied
-
-            // Call callback to apply changes
-            m_fontApplyCallback(m_currentConfig);
+            nk_layout_row_dynamic(m_ctx, 60, 1);
+            nk_label_colored(m_ctx, "No fonts found in /fonts directory", NK_TEXT_CENTERED, nk_rgb(255, 255, 0));
+            nk_label(m_ctx, "Please add .ttf or .otf files to the fonts folder", NK_TEXT_CENTERED);
         }
         else
         {
-            std::cout << "Error: Cannot apply - invalid font or no callback" << std::endl;
-        }
-    }
+            // Font Family Dropdown
+            nk_layout_row_dynamic(m_ctx, 20, 1);
+            nk_label(m_ctx, "Font Family:", NK_TEXT_LEFT);
 
-    if (!hasValidFont)
-    {
-        ImGui::EndDisabled();
-    }
-
-    ImGui::SameLine();
-
-    if (ImGui::Button("Close", ImVec2(90, 40)))
-    {
-        std::cout << "Close button clicked!" << std::endl;
-        m_showFontMenu = false;
-        // Reset temp config to current config
-        m_tempConfig = m_currentConfig;
-        m_selectedFontIndex = findFontIndex(m_currentConfig.fontName);
-
-        // Reset reading style index
-        auto allStyles = OptionsManager::getAllReadingStyles();
-        m_selectedStyleIndex = 0;
-        for (size_t i = 0; i < allStyles.size(); i++)
-        {
-            if (allStyles[i] == m_currentConfig.readingStyle)
+            // Ensure selected index is valid
+            if (m_selectedFontIndex < 0 || m_selectedFontIndex >= (int) fonts.size())
             {
-                m_selectedStyleIndex = static_cast<int>(i);
-                break;
+                m_selectedFontIndex = 0;
+            }
+
+            // Create dropdown
+            nk_layout_row_dynamic(m_ctx, 25, 1);
+
+            // Highlight font dropdown if focused
+            if (m_mainScreenFocusIndex == WIDGET_FONT_DROPDOWN)
+            {
+                m_ctx->style.combo.button.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+                m_ctx->style.combo.button.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+                m_ctx->style.combo.button.active = nk_style_item_color(nk_rgb(0, 102, 235));
+                m_ctx->style.combo.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+                m_ctx->style.combo.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+                m_ctx->style.combo.active = nk_style_item_color(nk_rgb(0, 102, 235));
+            }
+            const char* currentFont = fonts[m_selectedFontIndex].displayName.c_str();
+
+            if (m_fontDropdownHighlightedIndex < 0 || m_fontDropdownHighlightedIndex >= (int) fonts.size())
+            {
+                m_fontDropdownHighlightedIndex = (m_selectedFontIndex >= 0 && m_selectedFontIndex < (int) fonts.size()) ? m_selectedFontIndex : 0;
+            }
+
+            bool forceOpen = m_fontDropdownOpen || m_fontDropdownSelectRequested || m_fontDropdownCancelRequested;
+            if (nk_combo_begin_label_controller(m_ctx, currentFont, nk_vec2(nk_widget_width(m_ctx), 200), forceOpen))
+            {
+                nk_layout_row_dynamic(m_ctx, 20, 1);
+                for (size_t i = 0; i < fonts.size(); ++i)
+                {
+                    m_ctx->style.selectable = originalSelectableStyle;
+
+                    bool isHighlighted = m_fontDropdownOpen && (m_fontDropdownHighlightedIndex == static_cast<int>(i));
+                    if (isHighlighted)
+                    {
+                        m_ctx->style.selectable.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+                        m_ctx->style.selectable.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+                        m_ctx->style.selectable.pressed = nk_style_item_color(nk_rgb(0, 102, 235));
+                        m_ctx->style.selectable.text_normal = nk_rgb(255, 255, 255);
+                        m_ctx->style.selectable.text_hover = nk_rgb(255, 255, 255);
+                        m_ctx->style.selectable.text_pressed = nk_rgb(255, 255, 255);
+                    }
+
+                    nk_bool isSelected = (m_selectedFontIndex == static_cast<int>(i));
+                    nk_bool selectionChanged = nk_selectable_label(m_ctx, fonts[i].displayName.c_str(), NK_TEXT_LEFT, &isSelected);
+                    if (selectionChanged && isSelected)
+                    {
+                        std::cout << "[DEBUG] Font selected: " << fonts[i].displayName << std::endl;
+                        m_selectedFontIndex = static_cast<int>(i);
+                        m_tempConfig.fontPath = fonts[i].filePath;
+                        m_tempConfig.fontName = fonts[i].displayName;
+                        m_fontDropdownHighlightedIndex = m_selectedFontIndex;
+                        nk_combo_close(m_ctx);
+                        m_fontDropdownOpen = false;
+                        m_fontDropdownSelectRequested = false;
+                        m_fontDropdownCancelRequested = false;
+                    }
+                }
+
+                if (m_fontDropdownSelectRequested)
+                {
+                    if (m_fontDropdownHighlightedIndex >= 0 && m_fontDropdownHighlightedIndex < (int) fonts.size())
+                    {
+                        int chosenIndex = m_fontDropdownHighlightedIndex;
+                        std::cout << "[DEBUG] Font selected (controller): " << fonts[chosenIndex].displayName << std::endl;
+                        m_selectedFontIndex = chosenIndex;
+                        m_tempConfig.fontPath = fonts[chosenIndex].filePath;
+                        m_tempConfig.fontName = fonts[chosenIndex].displayName;
+                    }
+                    nk_combo_close(m_ctx);
+                    m_fontDropdownOpen = false;
+                    m_fontDropdownSelectRequested = false;
+                    m_fontDropdownHighlightedIndex = m_selectedFontIndex;
+                }
+
+                if (m_fontDropdownCancelRequested)
+                {
+                    nk_combo_close(m_ctx);
+                    m_fontDropdownOpen = false;
+                    m_fontDropdownCancelRequested = false;
+                    m_fontDropdownHighlightedIndex = m_selectedFontIndex;
+                }
+
+                nk_combo_end(m_ctx);
+                m_ctx->style.selectable = originalSelectableStyle;
+            }
+            else
+            {
+                if (m_fontDropdownOpen || m_fontDropdownSelectRequested || m_fontDropdownCancelRequested)
+                {
+                    m_fontDropdownOpen = false;
+                    m_fontDropdownSelectRequested = false;
+                    m_fontDropdownCancelRequested = false;
+                    m_fontDropdownHighlightedIndex = m_selectedFontIndex;
+                }
+                m_ctx->style.selectable = originalSelectableStyle;
+            }
+
+            // Restore combo style
+            m_ctx->style.combo = originalComboStyle;
+            nk_layout_row_dynamic(m_ctx, 10, 1); // Spacing
+
+            // Font Size Section
+            nk_layout_row_dynamic(m_ctx, 20, 1);
+            nk_label(m_ctx, "Font Size (pt):", NK_TEXT_LEFT);
+
+            // Font size input field
+            nk_layout_row_dynamic(m_ctx, 25, 1);
+
+            // Highlight font size input if focused
+            if (m_mainScreenFocusIndex == WIDGET_FONT_SIZE_INPUT)
+            {
+                m_ctx->style.edit.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+                m_ctx->style.edit.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+                m_ctx->style.edit.active = nk_style_item_color(nk_rgb(0, 102, 235));
+            }
+
+            int editFlags = nk_edit_string_zero_terminated(m_ctx, NK_EDIT_FIELD | NK_EDIT_SIG_ENTER | NK_EDIT_SELECTABLE,
+                                                           m_fontSizeInput, sizeof(m_fontSizeInput), nk_filter_decimal);
+
+            // Restore edit style
+            m_ctx->style.edit = originalEditStyle;
+            if (editFlags & NK_EDIT_ACTIVATED)
+            {
+                std::cout << "[DEBUG] Font size edit field ACTIVATED" << std::endl;
+            }
+            if (editFlags & NK_EDIT_DEACTIVATED)
+            {
+                std::cout << "[DEBUG] Font size edit field DEACTIVATED" << std::endl;
+            }
+            if (editFlags & NK_EDIT_COMMITED)
+            {
+                std::cout << "[DEBUG] Font size edit field COMMITTED: " << m_fontSizeInput << std::endl;
+                int newSize = std::atoi(m_fontSizeInput);
+                if (newSize >= 8 && newSize <= 72)
+                {
+                    m_tempConfig.fontSize = newSize;
+                }
+            }
+            // Also update if the value changes during editing
+            if (editFlags & NK_EDIT_ACTIVATED || editFlags & NK_EDIT_DEACTIVATED)
+            {
+                int newSize = std::atoi(m_fontSizeInput);
+                if (newSize >= 8 && newSize <= 72)
+                {
+                    m_tempConfig.fontSize = newSize;
+                }
+            }
+
+            // Font size slider
+            nk_layout_row_dynamic(m_ctx, 20, 1);
+
+            // Highlight font size slider if focused
+            if (m_mainScreenFocusIndex == WIDGET_FONT_SIZE_SLIDER)
+            {
+                m_ctx->style.slider.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+                m_ctx->style.slider.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+                m_ctx->style.slider.active = nk_style_item_color(nk_rgb(0, 102, 235));
+                m_ctx->style.slider.cursor_normal = nk_style_item_color(nk_rgb(0, 122, 255));
+                m_ctx->style.slider.cursor_hover = nk_style_item_color(nk_rgb(30, 142, 255));
+                m_ctx->style.slider.cursor_active = nk_style_item_color(nk_rgb(0, 102, 235));
+            }
+            float fontSize = static_cast<float>(m_tempConfig.fontSize);
+            if (nk_slider_float(m_ctx, 8.0f, &fontSize, 72.0f, 1.0f))
+            {
+                m_tempConfig.fontSize = static_cast<int>(fontSize);
+                snprintf(m_fontSizeInput, sizeof(m_fontSizeInput), "%d", m_tempConfig.fontSize);
+            }
+
+            // Restore slider style
+            m_ctx->style.slider = originalSliderStyle;
+
+            nk_layout_row_dynamic(m_ctx, 15, 1); // Spacing
+        }
+
+        // === ZOOM SETTINGS SECTION ===
+        nk_layout_row_dynamic(m_ctx, 25, 1);
+        nk_label(m_ctx, "Zoom Settings", NK_TEXT_LEFT);
+
+        // Separator line
+        nk_layout_row_dynamic(m_ctx, 1, 1);
+        bounds = nk_widget_bounds(m_ctx);
+        canvas = nk_window_get_canvas(m_ctx);
+        nk_stroke_line(canvas, bounds.x, bounds.y, bounds.x + bounds.w, bounds.y, 1.0f, nk_rgb(100, 100, 100));
+
+        nk_layout_row_dynamic(m_ctx, 20, 1);
+        nk_label(m_ctx, "Zoom Step (%) - Amount to zoom in/out with +/- keys:", NK_TEXT_LEFT);
+
+        // Zoom step input
+        nk_layout_row_dynamic(m_ctx, 25, 1);
+
+        // Highlight zoom step input if focused
+        if (m_mainScreenFocusIndex == WIDGET_ZOOM_STEP_INPUT)
+        {
+            m_ctx->style.edit.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+            m_ctx->style.edit.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+            m_ctx->style.edit.active = nk_style_item_color(nk_rgb(0, 102, 235));
+        }
+
+        int zoomEditFlags = nk_edit_string_zero_terminated(m_ctx, NK_EDIT_FIELD | NK_EDIT_SIG_ENTER | NK_EDIT_SELECTABLE,
+                                                           m_zoomStepInput, sizeof(m_zoomStepInput), nk_filter_decimal);
+
+        // Restore edit style
+        m_ctx->style.edit = originalEditStyle;
+        if (zoomEditFlags & NK_EDIT_COMMITED)
+        {
+            int newStep = std::atoi(m_zoomStepInput);
+            if (newStep >= 1 && newStep <= 50)
+            {
+                m_tempConfig.zoomStep = newStep;
+            }
+        }
+        // Also update if the value changes during editing
+        if (zoomEditFlags & NK_EDIT_ACTIVATED || zoomEditFlags & NK_EDIT_DEACTIVATED)
+        {
+            int newStep = std::atoi(m_zoomStepInput);
+            if (newStep >= 1 && newStep <= 50)
+            {
+                m_tempConfig.zoomStep = newStep;
             }
         }
 
-        // Trigger redraw to clear menu from screen
-        if (m_closeCallback)
+        // Zoom step slider
+        nk_layout_row_dynamic(m_ctx, 20, 1);
+
+        // Highlight zoom step slider if focused
+        if (m_mainScreenFocusIndex == WIDGET_ZOOM_STEP_SLIDER)
         {
-            m_closeCallback();
-        }
-    }
-
-    ImGui::SameLine();
-
-    if (ImGui::Button("Reset to Default", ImVec2(230, 40)))
-    {
-        // Reset to default config
-        m_tempConfig = FontConfig();
-        m_selectedFontIndex = 0;
-        m_selectedStyleIndex = 0; // Reset to Default style
-
-        // Set first available font as default
-        if (!fonts.empty())
-        {
-            m_tempConfig.fontPath = fonts[0].filePath;
-            m_tempConfig.fontName = fonts[0].displayName;
+            m_ctx->style.slider.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+            m_ctx->style.slider.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+            m_ctx->style.slider.active = nk_style_item_color(nk_rgb(0, 102, 235));
+            m_ctx->style.slider.cursor_normal = nk_style_item_color(nk_rgb(0, 122, 255));
+            m_ctx->style.slider.cursor_hover = nk_style_item_color(nk_rgb(30, 142, 255));
+            m_ctx->style.slider.cursor_active = nk_style_item_color(nk_rgb(0, 102, 235));
         }
 
-        // Reset page jump input
-        strcpy(m_pageJumpInput, "1");
-    }
+        float zoomStep = static_cast<float>(m_tempConfig.zoomStep);
+        if (nk_slider_float(m_ctx, 1.0f, &zoomStep, 50.0f, 1.0f))
+        {
+            m_tempConfig.zoomStep = static_cast<int>(zoomStep);
+            snprintf(m_zoomStepInput, sizeof(m_zoomStepInput), "%d", m_tempConfig.zoomStep);
+        }
 
-    ImGui::End();
+        // Restore zoom slider style
+        m_ctx->style.slider = originalSliderStyle;
+
+        nk_layout_row_dynamic(m_ctx, 15, 1); // Spacing
+
+        // === PAGE NAVIGATION SECTION ===
+        nk_layout_row_dynamic(m_ctx, 25, 1);
+        nk_label(m_ctx, "Page Navigation", NK_TEXT_LEFT);
+
+        // Separator line
+        nk_layout_row_dynamic(m_ctx, 1, 1);
+        bounds = nk_widget_bounds(m_ctx);
+        canvas = nk_window_get_canvas(m_ctx);
+        nk_stroke_line(canvas, bounds.x, bounds.y, bounds.x + bounds.w, bounds.y, 1.0f, nk_rgb(100, 100, 100));
+
+        // Current page display
+        char pageInfo[64];
+        snprintf(pageInfo, sizeof(pageInfo), "Current Page: %d / %d", m_currentPage + 1, m_pageCount);
+        nk_layout_row_dynamic(m_ctx, 20, 1);
+        nk_label(m_ctx, pageInfo, NK_TEXT_LEFT);
+
+        nk_layout_row_dynamic(m_ctx, 20, 1);
+        nk_label(m_ctx, "Jump to Page:", NK_TEXT_LEFT);
+
+        // Page jump input and buttons
+        nk_layout_row_template_begin(m_ctx, 25);
+        nk_layout_row_template_push_static(m_ctx, 100);
+        nk_layout_row_template_push_static(m_ctx, 60);
+        nk_layout_row_template_push_static(m_ctx, 100);
+        nk_layout_row_template_end(m_ctx);
+
+        // Highlight page jump input if focused
+        if (m_mainScreenFocusIndex == WIDGET_PAGE_JUMP_INPUT)
+        {
+            m_ctx->style.edit.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+            m_ctx->style.edit.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+            m_ctx->style.edit.active = nk_style_item_color(nk_rgb(0, 102, 235));
+        }
+
+        nk_edit_string_zero_terminated(m_ctx, NK_EDIT_FIELD | NK_EDIT_SELECTABLE, m_pageJumpInput, sizeof(m_pageJumpInput), nk_filter_decimal);
+
+        // Restore edit style
+        m_ctx->style.edit = originalEditStyle;
+
+        // Validate page input
+        bool validPageInput = false;
+        int targetPage = std::atoi(m_pageJumpInput);
+        if (targetPage >= 1 && targetPage <= m_pageCount)
+        {
+            validPageInput = true;
+        }
+
+        // Go button
+        if (!validPageInput)
+        {
+            nk_widget_disable_begin(m_ctx);
+        }
+
+        // Highlight Go button if focused
+        if (m_mainScreenFocusIndex == WIDGET_GO_BUTTON)
+        {
+            m_ctx->style.button.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+            m_ctx->style.button.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+            m_ctx->style.button.active = nk_style_item_color(nk_rgb(0, 102, 235));
+        }
+
+        if (nk_button_label(m_ctx, "Go"))
+        {
+            if (validPageInput && m_pageJumpCallback)
+            {
+                m_pageJumpCallback(targetPage - 1); // Convert to 0-based
+            }
+        }
+        if (!validPageInput)
+        {
+            nk_widget_disable_end(m_ctx);
+        }
+
+        // Restore Go button style
+        m_ctx->style.button = originalButtonStyle;
+
+        // Highlight Number Pad button if focused
+        if (m_mainScreenFocusIndex == WIDGET_NUMPAD_BUTTON)
+        {
+            m_ctx->style.button.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+            m_ctx->style.button.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+            m_ctx->style.button.active = nk_style_item_color(nk_rgb(0, 102, 235));
+        }
+
+        // Number pad button
+        if (nk_button_label(m_ctx, "Number Pad"))
+        {
+            showNumberPad();
+        }
+
+        // Restore Number Pad button style
+        m_ctx->style.button = originalButtonStyle;
+
+        // Show validation message if needed
+        if (!validPageInput && targetPage != 0)
+        {
+            nk_layout_row_dynamic(m_ctx, 20, 1);
+            nk_label_colored(m_ctx, "Invalid page number", NK_TEXT_LEFT, nk_rgb(255, 100, 100));
+        }
+
+        nk_layout_row_dynamic(m_ctx, 15, 1); // Spacing
+
+        // === CONTROLLER HINTS ===
+        nk_layout_row_dynamic(m_ctx, 20, 1);
+        nk_label_colored(m_ctx, "Controller: D-Pad=Navigate, A=Select, B=Close/Unfocus, Y=Apply, X=Reset", NK_TEXT_CENTERED, nk_rgb(150, 150, 150));
+        nk_layout_row_dynamic(m_ctx, 20, 1);
+        nk_label_colored(m_ctx, "L/R Shoulder=Tab Between Fields, Start=NumberPad", NK_TEXT_CENTERED, nk_rgb(150, 150, 150));
+
+        // === BUTTONS SECTION ===
+        // Separator line
+        nk_layout_row_dynamic(m_ctx, 1, 1);
+        bounds = nk_widget_bounds(m_ctx);
+        canvas = nk_window_get_canvas(m_ctx);
+        nk_stroke_line(canvas, bounds.x, bounds.y, bounds.x + bounds.w, bounds.y, 1.0f, nk_rgb(100, 100, 100));
+
+        nk_layout_row_dynamic(m_ctx, 10, 1); // Spacing
+
+        bool hasValidFont = !fonts.empty() && m_selectedFontIndex >= 0 && m_selectedFontIndex < (int) fonts.size();
+
+        nk_layout_row_template_begin(m_ctx, 30);
+        nk_layout_row_template_push_static(m_ctx, 90);
+        nk_layout_row_template_push_static(m_ctx, 20);
+        nk_layout_row_template_push_static(m_ctx, 90);
+        nk_layout_row_template_push_static(m_ctx, 20);
+        nk_layout_row_template_push_static(m_ctx, 110);
+        nk_layout_row_template_end(m_ctx);
+
+        // Apply button
+        if (!hasValidFont)
+        {
+            nk_widget_disable_begin(m_ctx);
+        }
+
+        // Highlight Apply button if focused
+        if (m_mainScreenFocusIndex == WIDGET_APPLY_BUTTON)
+        {
+            m_ctx->style.button.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+            m_ctx->style.button.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+            m_ctx->style.button.active = nk_style_item_color(nk_rgb(0, 102, 235));
+        }
+
+        if (nk_button_label(m_ctx, "Apply"))
+        {
+            if (hasValidFont && m_fontApplyCallback)
+            {
+                // Update current config
+                m_currentConfig = m_tempConfig;
+
+                // Save config to file
+                m_optionsManager.saveConfig(m_currentConfig);
+
+                // Call callback to apply changes
+                m_fontApplyCallback(m_currentConfig);
+            }
+        }
+
+        // Restore Apply button style
+        m_ctx->style.button = originalButtonStyle;
+        if (!hasValidFont)
+        {
+            nk_widget_disable_end(m_ctx);
+        }
+
+        nk_spacing(m_ctx, 1); // Empty column
+
+        // Close button
+        if (nk_button_label(m_ctx, "Close"))
+        {
+            m_showFontMenu = false;
+            m_fontDropdownOpen = false;
+            m_fontDropdownSelectRequested = false;
+            m_fontDropdownCancelRequested = false;
+            // Reset temp config to current config
+            m_tempConfig = m_currentConfig;
+            m_selectedFontIndex = findFontIndex(m_currentConfig.fontName);
+            m_fontDropdownHighlightedIndex = m_selectedFontIndex;
+
+            // Reset input fields
+            snprintf(m_fontSizeInput, sizeof(m_fontSizeInput), "%d", m_currentConfig.fontSize);
+            snprintf(m_zoomStepInput, sizeof(m_zoomStepInput), "%d", m_currentConfig.zoomStep);
+
+            if (m_closeCallback)
+            {
+                m_closeCallback();
+            }
+        }
+
+        nk_spacing(m_ctx, 1); // Empty column
+
+        // Reset to Default button
+        if (nk_button_label(m_ctx, "Reset to Default"))
+        {
+            // Reset to default config
+            m_tempConfig = FontConfig();
+            m_selectedFontIndex = 0;
+            m_fontDropdownHighlightedIndex = m_selectedFontIndex;
+
+            // Reset input fields to defaults
+            strcpy(m_fontSizeInput, "12");
+            strcpy(m_zoomStepInput, "10");
+            strcpy(m_pageJumpInput, "1");
+        }
+    }
+    nk_end(m_ctx);
 }
 
 void GuiManager::renderNumberPad()
 {
-    // Center the number pad window
-    ImGuiIO& io = ImGui::GetIO();
-    ImVec2 center = ImVec2(io.DisplaySize.x * 0.5f, io.DisplaySize.y * 0.5f);
-
-    ImGui::SetNextWindowPos(center, ImGuiCond_Always, ImVec2(0.5f, 0.5f));
-    ImGui::SetNextWindowSize(ImVec2(310, 600), ImGuiCond_Always);
-
-    if (!ImGui::Begin("Number Pad", nullptr, ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoCollapse))
-    {
-        ImGui::End();
+    if (!m_ctx)
         return;
-    }
 
-    ImGui::Text("Enter Page Number:");
-    ImGui::Separator();
+    int windowWidth, windowHeight;
+    SDL_GetWindowSize(m_window, &windowWidth, &windowHeight);
 
-    // Display current input
-    ImGui::Text("Page: %s", m_pageJumpInput);
-    ImGui::Separator();
+    // Center the number pad window
+    float centerX = windowWidth * 0.5f;
+    float centerY = windowHeight * 0.5f;
+    float windowW = 300.0f;
+    float windowH = 400.0f;
 
-    // Number pad layout (5x3 grid)
-    // Row 0: 7, 8, 9
-    // Row 1: 4, 5, 6
-    // Row 2: 1, 2, 3
-    // Row 3: Clear, 0, Back
-    // Row 4: Go, Cancel, (empty)
-
-    const char* buttons[5][3] = {
-        {"7", "8", "9"},        // Row 0: numbers
-        {"4", "5", "6"},        // Row 1: numbers
-        {"1", "2", "3"},        // Row 2: numbers
-        {"Clear", "0", "Back"}, // Row 3: utility buttons
-        {"Go", "Cancel", ""}    // Row 4: action buttons
-    };
-
-    const float buttonSize = 85.0f;
-
-    for (int row = 0; row < 5; row++)
+    if (nk_begin(m_ctx, "Number Pad", nk_rect(centerX - windowW / 2, centerY - windowH / 2, windowW, windowH),
+                 NK_WINDOW_BORDER | NK_WINDOW_TITLE | NK_WINDOW_NO_SCROLLBAR))
     {
-        for (int col = 0; col < 3; col++)
+        nk_layout_row_dynamic(m_ctx, 20, 1);
+        nk_label(m_ctx, "Enter Page Number:", NK_TEXT_CENTERED);
+
+        // Separator line
+        nk_layout_row_dynamic(m_ctx, 1, 1);
+        struct nk_rect bounds = nk_widget_bounds(m_ctx);
+        struct nk_command_buffer* canvas = nk_window_get_canvas(m_ctx);
+        nk_stroke_line(canvas, bounds.x, bounds.y, bounds.x + bounds.w, bounds.y, 1.0f, nk_rgb(100, 100, 100));
+
+        // Display current input
+        char displayText[32];
+        snprintf(displayText, sizeof(displayText), "Page: %s", strlen(m_pageJumpInput) > 0 ? m_pageJumpInput : "_");
+        nk_layout_row_dynamic(m_ctx, 30, 1);
+        nk_label(m_ctx, displayText, NK_TEXT_CENTERED);
+
+        // Separator line
+        nk_layout_row_dynamic(m_ctx, 1, 1);
+        bounds = nk_widget_bounds(m_ctx);
+        canvas = nk_window_get_canvas(m_ctx);
+        nk_stroke_line(canvas, bounds.x, bounds.y, bounds.x + bounds.w, bounds.y, 1.0f, nk_rgb(100, 100, 100));
+
+        // Number pad layout (3x4 grid)
+        // Row 0: 7, 8, 9
+        // Row 1: 4, 5, 6
+        // Row 2: 1, 2, 3
+        // Row 3: Clear, 0, Back
+
+        const char* numberGrid[4][3] = {
+            {"7", "8", "9"},
+            {"4", "5", "6"},
+            {"1", "2", "3"},
+            {"Clear", "0", "Back"}};
+
+        for (int row = 0; row < 4; row++)
         {
-            // Skip empty buttons
-            if (strlen(buttons[row][col]) == 0)
-                continue;
-
-            if (col > 0 && strlen(buttons[row][col - 1]) > 0)
-                ImGui::SameLine();
-
-            // Highlight selected button
-            bool isSelected = (m_numberPadSelectedRow == row && m_numberPadSelectedCol == col);
-            if (isSelected)
+            nk_layout_row_dynamic(m_ctx, 50, 3);
+            for (int col = 0; col < 3; col++)
             {
-                ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0.0f, 0.5f, 1.0f, 1.0f)); // Blue highlight
-                ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4(0.0f, 0.6f, 1.0f, 1.0f));
-                ImGui::PushStyleColor(ImGuiCol_ButtonActive, ImVec4(0.0f, 0.4f, 0.8f, 1.0f));
-            }
+                const char* buttonText = numberGrid[row][col];
 
-            // Disable button to prevent ImGui from handling input
-            // All input is handled by controller in handleNumberPadInput
-            ImGui::PushItemFlag(ImGuiItemFlags_Disabled, true);
-            ImGui::PushStyleVar(ImGuiStyleVar_Alpha, 1.0f); // Keep full opacity even when disabled
+                // Highlight selected button for controller navigation
+                bool isSelected = (m_numberPadSelectedRow == row && m_numberPadSelectedCol == col);
 
-            ImGui::Button(buttons[row][col], ImVec2(buttonSize, buttonSize));
+                struct nk_style_button oldStyle = m_ctx->style.button;
+                if (isSelected)
+                {
+                    m_ctx->style.button.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+                    m_ctx->style.button.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+                    m_ctx->style.button.active = nk_style_item_color(nk_rgb(0, 102, 235));
+                }
 
-            ImGui::PopStyleVar();
-            ImGui::PopItemFlag();
+                if (nk_button_label(m_ctx, buttonText))
+                {
+                    if (strcmp(buttonText, "Clear") == 0)
+                    {
+                        // Clear all input
+                        strcpy(m_pageJumpInput, "");
+                    }
+                    else if (strcmp(buttonText, "Back") == 0)
+                    {
+                        // Backspace - remove last character
+                        int len = strlen(m_pageJumpInput);
+                        if (len > 0)
+                        {
+                            m_pageJumpInput[len - 1] = '\0';
+                        }
+                    }
+                    else
+                    {
+                        // Add digit if there's space
+                        int len = strlen(m_pageJumpInput);
+                        if (len < (int) sizeof(m_pageJumpInput) - 1)
+                        {
+                            m_pageJumpInput[len] = buttonText[0];
+                            m_pageJumpInput[len + 1] = '\0';
+                        }
+                    }
+                }
 
-            if (isSelected)
-            {
-                ImGui::PopStyleColor(3);
+                // Restore original button style
+                if (isSelected)
+                {
+                    m_ctx->style.button = oldStyle;
+                }
             }
         }
-    }
 
-    // Show validation message if needed
-    int targetPage = strlen(m_pageJumpInput) > 0 ? std::atoi(m_pageJumpInput) : 0;
-    if (targetPage != 0 && (targetPage < 1 || targetPage > m_pageCount))
-    {
-        ImGui::TextColored(ImVec4(1, 0.4f, 0.4f, 1), "Invalid page number");
-    }
+        nk_layout_row_dynamic(m_ctx, 10, 1); // Spacing
 
-    ImGui::End();
+        // Action buttons
+        nk_layout_row_dynamic(m_ctx, 40, 2);
+
+        // Go button
+        int targetPage = strlen(m_pageJumpInput) > 0 ? std::atoi(m_pageJumpInput) : 0;
+        bool validPage = (targetPage >= 1 && targetPage <= m_pageCount);
+
+        // Highlight Go button if selected
+        bool goSelected = (m_numberPadSelectedRow == 4 && m_numberPadSelectedCol == 0);
+        struct nk_style_button oldStyleGo = m_ctx->style.button;
+        if (goSelected)
+        {
+            m_ctx->style.button.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+            m_ctx->style.button.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+            m_ctx->style.button.active = nk_style_item_color(nk_rgb(0, 102, 235));
+        }
+
+        if (!validPage)
+        {
+            nk_widget_disable_begin(m_ctx);
+        }
+        if (nk_button_label(m_ctx, "Go"))
+        {
+            if (validPage && m_pageJumpCallback)
+            {
+                m_pageJumpCallback(targetPage - 1); // Convert to 0-based
+                hideNumberPad();
+            }
+        }
+        if (!validPage)
+        {
+            nk_widget_disable_end(m_ctx);
+        }
+
+        // Restore button style
+        if (goSelected)
+        {
+            m_ctx->style.button = oldStyleGo;
+        }
+
+        // Highlight Cancel button if selected
+        bool cancelSelected = (m_numberPadSelectedRow == 4 && m_numberPadSelectedCol == 1);
+        if (cancelSelected)
+        {
+            m_ctx->style.button.normal = nk_style_item_color(nk_rgb(0, 122, 255));
+            m_ctx->style.button.hover = nk_style_item_color(nk_rgb(30, 142, 255));
+            m_ctx->style.button.active = nk_style_item_color(nk_rgb(0, 102, 235));
+        }
+
+        // Cancel button
+        if (nk_button_label(m_ctx, "Cancel"))
+        {
+            hideNumberPad();
+        }
+
+        // Restore button style
+        if (cancelSelected)
+        {
+            m_ctx->style.button = oldStyleGo;
+        }
+
+        // Show validation message if needed
+        if (targetPage != 0 && !validPage)
+        {
+            nk_layout_row_dynamic(m_ctx, 20, 1);
+            nk_label_colored(m_ctx, "Invalid page number", NK_TEXT_CENTERED, nk_rgb(255, 100, 100));
+        }
+
+        // Controller hints
+        nk_layout_row_dynamic(m_ctx, 15, 1);
+        nk_label_colored(m_ctx, "Controller: D-Pad=Navigate, A=Select, B=Backspace, Start=Go", NK_TEXT_CENTERED, nk_rgb(150, 150, 150));
+    }
+    nk_end(m_ctx);
 }
 
 bool GuiManager::handleNumberPadInput(const SDL_Event& event)
@@ -914,6 +1111,22 @@ bool GuiManager::handleNumberPadInput(const SDL_Event& event)
     if (!m_showNumberPad)
     {
         return false;
+    }
+
+    if (event.type == SDL_JOYBUTTONDOWN)
+    {
+        if (event.jbutton.button == 10)
+        {
+            if (m_showNumberPad)
+            {
+                hideNumberPad();
+            }
+            else if (m_showFontMenu)
+            {
+                closeFontMenu();
+            }
+            return true;
+        }
     }
 
     if (event.type == SDL_CONTROLLERBUTTONDOWN)
@@ -926,56 +1139,17 @@ bool GuiManager::handleNumberPadInput(const SDL_Event& event)
         }
         m_lastButtonPressTime = currentTime;
 
-        // Map physical button to logical button
-        SDL_GameControllerButton physicalButton = static_cast<SDL_GameControllerButton>(event.cbutton.button);
-        LogicalButton logicalButton = LogicalButton::Accept; // Default
-
-#ifdef TG5040_PLATFORM
-        // On TG5040, use direct SDL button mapping (not ButtonMapper) to avoid interference with ImGui patch
-        // SDL reports buttons swapped: Physical A = SDL BUTTON_B, Physical B = SDL BUTTON_A
-        // For number pad, we want:
-        //   - Physical A (SDL BUTTON_B) = Accept (activate selected number pad button)
-        //   - Physical B (SDL BUTTON_A) = Already handled above to close number pad
-        switch (physicalButton)
+        switch (event.cbutton.button)
         {
-        case SDL_CONTROLLER_BUTTON_B:
-            logicalButton = LogicalButton::Accept; // Physical A
-            break;
         case SDL_CONTROLLER_BUTTON_DPAD_UP:
-            logicalButton = LogicalButton::DPadUp;
-            break;
-        case SDL_CONTROLLER_BUTTON_DPAD_DOWN:
-            logicalButton = LogicalButton::DPadDown;
-            break;
-        case SDL_CONTROLLER_BUTTON_DPAD_LEFT:
-            logicalButton = LogicalButton::DPadLeft;
-            break;
-        case SDL_CONTROLLER_BUTTON_DPAD_RIGHT:
-            logicalButton = LogicalButton::DPadRight;
-            break;
-        default:
-            return true; // Ignore other buttons
-        }
-#else
-        // On desktop platforms, use ButtonMapper
-        if (m_buttonMapper)
-        {
-            logicalButton = m_buttonMapper->mapButton(physicalButton);
-        }
-#endif
-
-        switch (logicalButton)
-        {
-        case LogicalButton::DPadUp:
-            m_numberPadSelectedRow = (m_numberPadSelectedRow - 1 + 5) % 5;
+            m_numberPadSelectedRow = (m_numberPadSelectedRow - 1 + 5) % 5; // 5 rows (4 number rows + 1 action row)
             return true;
-        case LogicalButton::DPadDown:
+        case SDL_CONTROLLER_BUTTON_DPAD_DOWN:
             m_numberPadSelectedRow = (m_numberPadSelectedRow + 1) % 5;
             return true;
-        case LogicalButton::DPadLeft:
-            // Handle navigation - row 4 has only 2 buttons (Go, Cancel)
+        case SDL_CONTROLLER_BUTTON_DPAD_LEFT:
             if (m_numberPadSelectedRow == 4)
-            {
+            { // Action row (Go/Cancel)
                 m_numberPadSelectedCol = (m_numberPadSelectedCol - 1 + 2) % 2;
             }
             else
@@ -983,10 +1157,9 @@ bool GuiManager::handleNumberPadInput(const SDL_Event& event)
                 m_numberPadSelectedCol = (m_numberPadSelectedCol - 1 + 3) % 3;
             }
             return true;
-        case LogicalButton::DPadRight:
-            // Handle navigation - row 4 has only 2 buttons (Go, Cancel)
+        case SDL_CONTROLLER_BUTTON_DPAD_RIGHT:
             if (m_numberPadSelectedRow == 4)
-            {
+            { // Action row (Go/Cancel)
                 m_numberPadSelectedCol = (m_numberPadSelectedCol + 1) % 2;
             }
             else
@@ -994,66 +1167,62 @@ bool GuiManager::handleNumberPadInput(const SDL_Event& event)
                 m_numberPadSelectedCol = (m_numberPadSelectedCol + 1) % 3;
             }
             return true;
-        case LogicalButton::Accept:
+        case kAcceptButton:
         {
-            // Handle selected button press - use EXACT same layout as renderNumberPad
-            const char* buttons[5][3] = {
-                {"7", "8", "9"},        // Row 0: numbers
-                {"4", "5", "6"},        // Row 1: numbers
-                {"1", "2", "3"},        // Row 2: numbers
-                {"Clear", "0", "Back"}, // Row 3: utility buttons
-                {"Go", "Cancel", ""}    // Row 4: action buttons
-            };
-
-            const char* buttonText = buttons[m_numberPadSelectedRow][m_numberPadSelectedCol];
-
-            // Skip empty buttons
-            if (strlen(buttonText) == 0)
-            {
-                return true;
-            }
-
-            if (strcmp(buttonText, "Clear") == 0)
-            {
-                strcpy(m_pageJumpInput, "");
-            }
-            else if (strcmp(buttonText, "Back") == 0)
-            {
-                // Backspace - remove last character
-                int len = strlen(m_pageJumpInput);
-                if (len > 0)
-                {
-                    m_pageJumpInput[len - 1] = '\0';
+            // Handle selected button press
+            if (m_numberPadSelectedRow == 4)
+            { // Action row
+                if (m_numberPadSelectedCol == 0)
+                { // Go
+                    int targetPage = strlen(m_pageJumpInput) > 0 ? std::atoi(m_pageJumpInput) : 0;
+                    if (targetPage >= 1 && targetPage <= m_pageCount && m_pageJumpCallback)
+                    {
+                        m_pageJumpCallback(targetPage - 1);
+                        hideNumberPad();
+                    }
                 }
-            }
-            else if (strcmp(buttonText, "Go") == 0)
-            {
-                // Go to page if valid
-                int targetPage = strlen(m_pageJumpInput) > 0 ? std::atoi(m_pageJumpInput) : 0;
-                if (targetPage >= 1 && targetPage <= m_pageCount && m_pageJumpCallback)
-                {
-                    m_pageJumpCallback(targetPage - 1);
+                else
+                { // Cancel
                     hideNumberPad();
                 }
             }
-            else if (strcmp(buttonText, "Cancel") == 0)
-            {
-                // Cancel - close number pad without action
-                hideNumberPad();
-            }
             else
             {
-                int len = strlen(m_pageJumpInput);
-                if (len < (int) sizeof(m_pageJumpInput) - 1)
+                // Number grid
+                const char* numberGrid[4][3] = {
+                    {"7", "8", "9"},
+                    {"4", "5", "6"},
+                    {"1", "2", "3"},
+                    {"Clear", "0", "Back"}};
+
+                const char* buttonText = numberGrid[m_numberPadSelectedRow][m_numberPadSelectedCol];
+
+                if (strcmp(buttonText, "Clear") == 0)
                 {
-                    m_pageJumpInput[len] = buttonText[0];
-                    m_pageJumpInput[len + 1] = '\0';
+                    strcpy(m_pageJumpInput, "");
+                }
+                else if (strcmp(buttonText, "Back") == 0)
+                {
+                    int len = strlen(m_pageJumpInput);
+                    if (len > 0)
+                    {
+                        m_pageJumpInput[len - 1] = '\0';
+                    }
+                }
+                else
+                {
+                    int len = strlen(m_pageJumpInput);
+                    if (len < (int) sizeof(m_pageJumpInput) - 1)
+                    {
+                        m_pageJumpInput[len] = buttonText[0];
+                        m_pageJumpInput[len + 1] = '\0';
+                    }
                 }
             }
             return true;
         }
-        case LogicalButton::Cancel:
-            // Backspace function - remove last character, or cancel if empty
+        case kCancelButton:
+            // Backspace function or cancel
             {
                 int len = strlen(m_pageJumpInput);
                 if (len > 0)
@@ -1062,13 +1231,12 @@ bool GuiManager::handleNumberPadInput(const SDL_Event& event)
                 }
                 else
                 {
-                    // If input is empty, close number pad
                     hideNumberPad();
                 }
                 return true;
             }
-        case LogicalButton::Menu:
-            // Go to page if valid (alternative method)
+        case SDL_CONTROLLER_BUTTON_START:
+            // Go to page if valid
             {
                 int targetPage = strlen(m_pageJumpInput) > 0 ? std::atoi(m_pageJumpInput) : 0;
                 if (targetPage >= 1 && targetPage <= m_pageCount && m_pageJumpCallback)
@@ -1083,25 +1251,538 @@ bool GuiManager::handleNumberPadInput(const SDL_Event& event)
         }
     }
 
-    // Handle keyboard input as backup
+    return false;
+}
+
+void GuiManager::showNumberPad()
+{
+    m_showNumberPad = true;
+    m_pageJumpInput[0] = '\0';
+    // Reset selection to middle of number pad (5)
+    m_numberPadSelectedRow = 1; // Row with 4,5,6
+    m_numberPadSelectedCol = 1; // Middle column (5)
+    // Reset debounce timer
+    m_lastButtonPressTime = 0;
+}
+
+void GuiManager::hideNumberPad()
+{
+    m_showNumberPad = false;
+}
+
+int GuiManager::findFontIndex(const std::string& fontName) const
+{
+    auto fonts = m_optionsManager.getAvailableFonts();
+    for (size_t i = 0; i < fonts.size(); ++i)
+    {
+        if (fonts[i].displayName == fontName)
+        {
+            return static_cast<int>(i);
+        }
+    }
+    return 0; // Default to first font if not found
+}
+
+void GuiManager::setupColorScheme()
+{
+    if (!m_ctx)
+        return;
+
+    struct nk_color table[NK_COLOR_COUNT];
+
+    // Modern dark theme with blue accents
+    table[NK_COLOR_TEXT] = nk_rgb(210, 210, 210);
+    table[NK_COLOR_WINDOW] = nk_rgb(45, 45, 48);
+    table[NK_COLOR_HEADER] = nk_rgb(40, 40, 43);
+    table[NK_COLOR_BORDER] = nk_rgb(65, 65, 70);
+    table[NK_COLOR_BUTTON] = nk_rgb(60, 60, 65);
+    table[NK_COLOR_BUTTON_HOVER] = nk_rgb(70, 70, 75);
+    table[NK_COLOR_BUTTON_ACTIVE] = nk_rgb(50, 50, 55);
+    table[NK_COLOR_TOGGLE] = nk_rgb(80, 80, 85);
+    table[NK_COLOR_TOGGLE_HOVER] = nk_rgb(90, 90, 95);
+    table[NK_COLOR_TOGGLE_CURSOR] = nk_rgb(0, 122, 255);
+    table[NK_COLOR_SELECT] = nk_rgb(40, 40, 43);
+    table[NK_COLOR_SELECT_ACTIVE] = nk_rgb(0, 122, 255);
+    table[NK_COLOR_SLIDER] = nk_rgb(60, 60, 65);
+    table[NK_COLOR_SLIDER_CURSOR] = nk_rgb(0, 122, 255);
+    table[NK_COLOR_SLIDER_CURSOR_HOVER] = nk_rgb(30, 142, 255);
+    table[NK_COLOR_SLIDER_CURSOR_ACTIVE] = nk_rgb(0, 102, 235);
+    table[NK_COLOR_PROPERTY] = nk_rgb(60, 60, 65);
+    table[NK_COLOR_EDIT] = nk_rgb(50, 50, 55);
+    table[NK_COLOR_EDIT_CURSOR] = nk_rgb(0, 122, 255);
+    table[NK_COLOR_COMBO] = nk_rgb(60, 60, 65);
+    table[NK_COLOR_CHART] = nk_rgb(60, 60, 65);
+    table[NK_COLOR_CHART_COLOR] = nk_rgb(0, 122, 255);
+    table[NK_COLOR_CHART_COLOR_HIGHLIGHT] = nk_rgb(30, 142, 255);
+    table[NK_COLOR_SCROLLBAR] = nk_rgb(50, 50, 55);
+    table[NK_COLOR_SCROLLBAR_CURSOR] = nk_rgb(70, 70, 75);
+    table[NK_COLOR_SCROLLBAR_CURSOR_HOVER] = nk_rgb(80, 80, 85);
+    table[NK_COLOR_SCROLLBAR_CURSOR_ACTIVE] = nk_rgb(60, 60, 65);
+    table[NK_COLOR_TAB_HEADER] = nk_rgb(50, 50, 55);
+
+    nk_style_from_table(m_ctx, table);
+
+    // Fine-tune some specific elements
+    m_ctx->style.window.fixed_background = nk_style_item_color(nk_rgb(35, 35, 38));
+    m_ctx->style.window.border_color = nk_rgb(0, 122, 255);
+    m_ctx->style.window.border = 2.0f;
+    m_ctx->style.window.rounding = 4.0f;
+
+    // Button styling
+    m_ctx->style.button.rounding = 3.0f;
+    m_ctx->style.button.border = 1.0f;
+    m_ctx->style.button.border_color = nk_rgb(80, 80, 85);
+
+    // Edit field styling
+    m_ctx->style.edit.rounding = 3.0f;
+    m_ctx->style.edit.border = 1.0f;
+    m_ctx->style.edit.border_color = nk_rgb(80, 80, 85);
+
+    // Combo styling
+    m_ctx->style.combo.rounding = 3.0f;
+    m_ctx->style.combo.border = 1.0f;
+    m_ctx->style.combo.border_color = nk_rgb(80, 80, 85);
+}
+
+bool GuiManager::handleKeyboardNavigation(const SDL_Event& event)
+{
+    if (!m_showFontMenu && !m_showNumberPad)
+    {
+        return false; // Only handle keyboard input when GUI is visible
+    }
+
     if (event.type == SDL_KEYDOWN)
     {
+        // Simple time-based debouncing (same as controller)
+        Uint32 currentTime = SDL_GetTicks();
+        if (currentTime - m_lastButtonPressTime < BUTTON_DEBOUNCE_MS)
+        {
+            return true; // Ignore rapid repeated presses
+        }
+        m_lastButtonPressTime = currentTime;
+
+        if (m_fontDropdownOpen)
+        {
+            const auto& fonts = m_optionsManager.getAvailableFonts();
+            int fontCount = static_cast<int>(fonts.size());
+            if (fontCount == 0)
+            {
+                m_fontDropdownHighlightedIndex = 0;
+            }
+
+            switch (event.key.keysym.sym)
+            {
+            case SDLK_UP:
+                if (fontCount > 0)
+                {
+                    m_fontDropdownHighlightedIndex = (m_fontDropdownHighlightedIndex - 1 + fontCount) % fontCount;
+                    std::cout << "[DEBUG] Keyboard dropdown UP - highlight: " << m_fontDropdownHighlightedIndex << std::endl;
+                }
+                return true;
+            case SDLK_DOWN:
+                if (fontCount > 0)
+                {
+                    m_fontDropdownHighlightedIndex = (m_fontDropdownHighlightedIndex + 1) % fontCount;
+                    std::cout << "[DEBUG] Keyboard dropdown DOWN - highlight: " << m_fontDropdownHighlightedIndex << std::endl;
+                }
+                return true;
+            case SDLK_RETURN:
+            case SDLK_SPACE:
+                std::cout << "[DEBUG] Keyboard confirm dropdown index " << m_fontDropdownHighlightedIndex << std::endl;
+                m_fontDropdownSelectRequested = true;
+                return true;
+            case SDLK_ESCAPE:
+                std::cout << "[DEBUG] Keyboard cancel dropdown" << std::endl;
+                m_fontDropdownCancelRequested = true;
+                return true;
+            case SDLK_LEFT:
+            case SDLK_RIGHT:
+                return true;
+            default:
+                break;
+            }
+        }
+
         switch (event.key.keysym.sym)
         {
-        case SDLK_ESCAPE:
-            hideNumberPad();
+        case SDLK_UP:
+            // Navigate up - move to previous widget (same as numpad logic)
+            m_mainScreenFocusIndex = (m_mainScreenFocusIndex - 1 + WIDGET_COUNT) % WIDGET_COUNT;
+            return true;
+        case SDLK_DOWN:
+            // Navigate down - move to next widget (same as numpad logic)
+            m_mainScreenFocusIndex = (m_mainScreenFocusIndex + 1) % WIDGET_COUNT;
+            return true;
+        case SDLK_LEFT:
+            // Navigate left - adjust widget value if applicable
+            adjustFocusedWidget(-1);
+            return true;
+        case SDLK_RIGHT:
+            // Navigate right - adjust widget value if applicable
+            adjustFocusedWidget(1);
+            return true;
+        case SDLK_TAB:
+            if (event.key.keysym.mod & KMOD_SHIFT)
+            {
+                // Shift+Tab for previous widget
+                m_mainScreenFocusIndex = (m_mainScreenFocusIndex - 1 + WIDGET_COUNT) % WIDGET_COUNT;
+            }
+            else
+            {
+                // Tab for next widget
+                m_mainScreenFocusIndex = (m_mainScreenFocusIndex + 1) % WIDGET_COUNT;
+            }
             return true;
         case SDLK_RETURN:
-        case SDLK_KP_ENTER:
-        {
-            int targetPage = strlen(m_pageJumpInput) > 0 ? std::atoi(m_pageJumpInput) : 0;
-            if (targetPage >= 1 && targetPage <= m_pageCount && m_pageJumpCallback)
+        case SDLK_SPACE:
+            // Activate the focused widget (same as numpad A button logic)
+            activateFocusedWidget();
+            return true;
+        case SDLK_ESCAPE:
+            // Handle escape to close menu (same as controller B button behavior)
+            nk_input_key(m_ctx, NK_KEY_TEXT_END, 1);
+            nk_input_key(m_ctx, NK_KEY_TEXT_END, 0);
+
+            if (m_showFontMenu)
             {
-                m_pageJumpCallback(targetPage - 1);
+                m_showFontMenu = false;
+                m_fontDropdownOpen = false;
+                m_fontDropdownSelectRequested = false;
+                m_fontDropdownCancelRequested = false;
+                // Reset temp config to current config
+                m_tempConfig = m_currentConfig;
+                m_selectedFontIndex = findFontIndex(m_currentConfig.fontName);
+                m_fontDropdownHighlightedIndex = m_selectedFontIndex;
+
+                // Reset input fields
+                snprintf(m_fontSizeInput, sizeof(m_fontSizeInput), "%d", m_currentConfig.fontSize);
+                snprintf(m_zoomStepInput, sizeof(m_zoomStepInput), "%d", m_currentConfig.zoomStep);
+
+                if (m_closeCallback)
+                {
+                    m_closeCallback();
+                }
+            }
+            else if (m_showNumberPad)
+            {
                 hideNumberPad();
             }
             return true;
+        default:
+            break;
         }
+    }
+
+    return false; // Event not handled
+}
+
+void GuiManager::adjustFocusedWidget(int direction)
+{
+    std::cout << "[DEBUG] Adjusting widget " << m_mainScreenFocusIndex << " by " << direction << std::endl;
+    switch (m_mainScreenFocusIndex)
+    {
+    case WIDGET_FONT_SIZE_SLIDER:
+        // Adjust font size
+        m_tempConfig.fontSize = std::clamp(m_tempConfig.fontSize + direction, 8, 72);
+        snprintf(m_fontSizeInput, sizeof(m_fontSizeInput), "%d", m_tempConfig.fontSize);
+        std::cout << "[DEBUG] Font size adjusted to: " << m_tempConfig.fontSize << std::endl;
+        break;
+    case WIDGET_ZOOM_STEP_SLIDER:
+        // Adjust zoom step
+        m_tempConfig.zoomStep = std::clamp(m_tempConfig.zoomStep + direction, 1, 50);
+        snprintf(m_zoomStepInput, sizeof(m_zoomStepInput), "%d", m_tempConfig.zoomStep);
+        std::cout << "[DEBUG] Zoom step adjusted to: " << m_tempConfig.zoomStep << std::endl;
+        break;
+    default:
+        // Other widgets don't respond to left/right adjustment
+        std::cout << "[DEBUG] Widget " << m_mainScreenFocusIndex << " is not adjustable" << std::endl;
+        break;
+    }
+}
+
+void GuiManager::activateFocusedWidget()
+{
+    switch (m_mainScreenFocusIndex)
+    {
+    case WIDGET_FONT_DROPDOWN:
+        if (!m_fontDropdownOpen)
+        {
+            const auto& fonts = m_optionsManager.getAvailableFonts();
+            if (!fonts.empty())
+            {
+                std::cout << "[DEBUG] Opening font dropdown" << std::endl;
+                m_fontDropdownOpen = true;
+                m_fontDropdownSelectRequested = false;
+                m_fontDropdownCancelRequested = false;
+                if (m_selectedFontIndex >= 0 && m_selectedFontIndex < (int) fonts.size())
+                {
+                    m_fontDropdownHighlightedIndex = m_selectedFontIndex;
+                }
+                else
+                {
+                    m_fontDropdownHighlightedIndex = 0;
+                }
+            }
+        }
+        else
+        {
+            std::cout << "[DEBUG] Confirming selection from font dropdown" << std::endl;
+            m_fontDropdownSelectRequested = true;
+        }
+        break;
+    case WIDGET_GO_BUTTON:
+        // Activate Go button
+        if (strlen(m_pageJumpInput) > 0)
+        {
+            int targetPage = std::atoi(m_pageJumpInput);
+            if (targetPage >= 1 && targetPage <= m_pageCount && m_pageJumpCallback)
+            {
+                m_pageJumpCallback(targetPage - 1);
+            }
+        }
+        break;
+    case WIDGET_NUMPAD_BUTTON:
+        // Show number pad
+        showNumberPad();
+        break;
+    case WIDGET_APPLY_BUTTON:
+        // Apply settings
+        if (m_fontApplyCallback)
+        {
+            m_fontApplyCallback(m_tempConfig);
+            m_currentConfig = m_tempConfig;
+        }
+        break;
+    case WIDGET_RESET_BUTTON:
+        // Reset to defaults
+        m_tempConfig = FontConfig();
+        m_selectedFontIndex = 0;
+        snprintf(m_fontSizeInput, sizeof(m_fontSizeInput), "%d", m_tempConfig.fontSize);
+        snprintf(m_zoomStepInput, sizeof(m_zoomStepInput), "%d", m_tempConfig.zoomStep);
+        break;
+    case WIDGET_CLOSE_BUTTON:
+        // Close menu
+        m_showFontMenu = false;
+        m_fontDropdownOpen = false;
+        m_fontDropdownSelectRequested = false;
+        m_fontDropdownCancelRequested = false;
+        m_tempConfig = m_currentConfig;
+        m_selectedFontIndex = findFontIndex(m_currentConfig.fontName);
+        m_fontDropdownHighlightedIndex = m_selectedFontIndex;
+        snprintf(m_fontSizeInput, sizeof(m_fontSizeInput), "%d", m_currentConfig.fontSize);
+        snprintf(m_zoomStepInput, sizeof(m_zoomStepInput), "%d", m_currentConfig.zoomStep);
+        if (m_closeCallback)
+        {
+            m_closeCallback();
+        }
+        break;
+    default:
+        // Other widgets don't have simple activation
+        break;
+    }
+}
+
+bool GuiManager::handleControllerInput(const SDL_Event& event)
+{
+    if (!m_showFontMenu && !m_showNumberPad)
+    {
+        return false; // Only handle controller input when GUI is visible
+    }
+
+    if (event.type == SDL_CONTROLLERBUTTONDOWN)
+    {
+        // Simple time-based debouncing
+        Uint32 currentTime = SDL_GetTicks();
+        if (currentTime - m_lastButtonPressTime < BUTTON_DEBOUNCE_MS)
+        {
+            return true; // Ignore rapid repeated presses
+        }
+        m_lastButtonPressTime = currentTime;
+
+        if (m_fontDropdownOpen)
+        {
+            const auto& fonts = m_optionsManager.getAvailableFonts();
+            int fontCount = static_cast<int>(fonts.size());
+            if (fontCount == 0)
+            {
+                m_fontDropdownHighlightedIndex = 0;
+            }
+
+            switch (event.cbutton.button)
+            {
+            case SDL_CONTROLLER_BUTTON_DPAD_UP:
+                if (fontCount > 0)
+                {
+                    m_fontDropdownHighlightedIndex = (m_fontDropdownHighlightedIndex - 1 + fontCount) % fontCount;
+                    std::cout << "[DEBUG] Controller dropdown UP - highlight: " << m_fontDropdownHighlightedIndex << std::endl;
+                }
+                return true;
+            case SDL_CONTROLLER_BUTTON_DPAD_DOWN:
+                if (fontCount > 0)
+                {
+                    m_fontDropdownHighlightedIndex = (m_fontDropdownHighlightedIndex + 1) % fontCount;
+                    std::cout << "[DEBUG] Controller dropdown DOWN - highlight: " << m_fontDropdownHighlightedIndex << std::endl;
+                }
+                return true;
+            case SDL_CONTROLLER_BUTTON_DPAD_LEFT:
+            case SDL_CONTROLLER_BUTTON_DPAD_RIGHT:
+                // Ignore left/right while dropdown is open
+                return true;
+            case kAcceptButton:
+                std::cout << "[DEBUG] Controller A - confirm dropdown index " << m_fontDropdownHighlightedIndex << std::endl;
+                m_fontDropdownSelectRequested = true;
+                return true;
+            case kCancelButton:
+                std::cout << "[DEBUG] Controller B - cancel dropdown" << std::endl;
+                m_fontDropdownCancelRequested = true;
+                return true;
+            default:
+                break;
+            }
+        }
+
+        // Handle main screen navigation using our custom system
+        if (m_showFontMenu)
+        {
+            std::cout << "[DEBUG] Controller dpad input for font menu: " << event.cbutton.button << std::endl;
+            switch (event.cbutton.button)
+            {
+            case SDL_CONTROLLER_BUTTON_DPAD_UP:
+                // Navigate up - move to previous widget
+                m_mainScreenFocusIndex = (m_mainScreenFocusIndex - 1 + WIDGET_COUNT) % WIDGET_COUNT;
+                std::cout << "[DEBUG] Controller UP - new focus: " << m_mainScreenFocusIndex << std::endl;
+                return true;
+            case SDL_CONTROLLER_BUTTON_DPAD_DOWN:
+                // Navigate down - move to next widget
+                m_mainScreenFocusIndex = (m_mainScreenFocusIndex + 1) % WIDGET_COUNT;
+                std::cout << "[DEBUG] Controller DOWN - new focus: " << m_mainScreenFocusIndex << std::endl;
+                return true;
+            case SDL_CONTROLLER_BUTTON_DPAD_LEFT:
+                std::cout << "[DEBUG] Controller LEFT - adjusting widget" << std::endl;
+                adjustFocusedWidget(-1);
+                return true;
+            case SDL_CONTROLLER_BUTTON_DPAD_RIGHT:
+                std::cout << "[DEBUG] Controller RIGHT - adjusting widget" << std::endl;
+                adjustFocusedWidget(1);
+                return true;
+            case kAcceptButton:
+                std::cout << "[DEBUG] Controller A - activating widget " << m_mainScreenFocusIndex << std::endl;
+                activateFocusedWidget();
+                return true;
+            }
+        }
+
+        switch (event.cbutton.button)
+        {
+        case kCancelButton:
+            // Send escape key first to exit any text input, then close menu
+            nk_input_key(m_ctx, NK_KEY_TEXT_END, 1);
+            nk_input_key(m_ctx, NK_KEY_TEXT_END, 0);
+
+            if (m_showFontMenu)
+            {
+                m_showFontMenu = false;
+                m_fontDropdownOpen = false;
+                m_fontDropdownSelectRequested = false;
+                m_fontDropdownCancelRequested = false;
+                // Reset temp config to current config
+                m_tempConfig = m_currentConfig;
+                m_selectedFontIndex = findFontIndex(m_currentConfig.fontName);
+                m_fontDropdownHighlightedIndex = m_selectedFontIndex;
+
+                // Reset input fields
+                snprintf(m_fontSizeInput, sizeof(m_fontSizeInput), "%d", m_currentConfig.fontSize);
+                snprintf(m_zoomStepInput, sizeof(m_zoomStepInput), "%d", m_currentConfig.zoomStep);
+
+                if (m_closeCallback)
+                {
+                    m_closeCallback();
+                }
+                return true;
+            }
+            return false;
+        case SDL_CONTROLLER_BUTTON_Y:
+            // Apply current settings (like Y for confirm/apply)
+            if (m_showFontMenu)
+            {
+                const auto& fonts = m_optionsManager.getAvailableFonts();
+                bool hasValidFont = !fonts.empty() && m_selectedFontIndex >= 0 && m_selectedFontIndex < (int) fonts.size();
+
+                if (hasValidFont && m_fontApplyCallback)
+                {
+                    // Update current config
+                    m_currentConfig = m_tempConfig;
+
+                    // Save config to file
+                    m_optionsManager.saveConfig(m_currentConfig);
+
+                    // Call callback to apply changes
+                    m_fontApplyCallback(m_currentConfig);
+                }
+                return true;
+            }
+            return false;
+        case SDL_CONTROLLER_BUTTON_X:
+            // Reset to defaults
+            if (m_showFontMenu)
+            {
+                m_tempConfig = FontConfig();
+                m_selectedFontIndex = 0;
+
+                // Reset input fields to defaults
+                strcpy(m_fontSizeInput, "12");
+                strcpy(m_zoomStepInput, "10");
+                strcpy(m_pageJumpInput, "1");
+                return true;
+            }
+            return false;
+        case SDL_CONTROLLER_BUTTON_START:
+            // Show number pad if in font menu
+            if (m_showFontMenu)
+            {
+                showNumberPad();
+                return true;
+            }
+            return false;
+        case SDL_CONTROLLER_BUTTON_RIGHTSHOULDER:
+            // Tab to next control (useful for navigating between input fields)
+            nk_input_key(m_ctx, NK_KEY_TAB, 1);
+            nk_input_key(m_ctx, NK_KEY_TAB, 0);
+            return true;
+        case SDL_CONTROLLER_BUTTON_LEFTSHOULDER:
+            // Shift+Tab to previous control - simulate holding shift while pressing tab
+            nk_input_key(m_ctx, NK_KEY_SHIFT, 1);
+            nk_input_key(m_ctx, NK_KEY_TAB, 1);
+            nk_input_key(m_ctx, NK_KEY_TAB, 0);
+            nk_input_key(m_ctx, NK_KEY_SHIFT, 0);
+            return true;
+        default:
+            break;
+        }
+    }
+
+    // Handle button up events to clear key states
+    if (event.type == SDL_CONTROLLERBUTTONUP)
+    {
+        switch (event.cbutton.button)
+        {
+        case SDL_CONTROLLER_BUTTON_DPAD_UP:
+            nk_input_key(m_ctx, NK_KEY_UP, 0);
+            return true;
+        case SDL_CONTROLLER_BUTTON_DPAD_DOWN:
+            nk_input_key(m_ctx, NK_KEY_DOWN, 0);
+            return true;
+        case SDL_CONTROLLER_BUTTON_DPAD_LEFT:
+            nk_input_key(m_ctx, NK_KEY_LEFT, 0);
+            return true;
+        case SDL_CONTROLLER_BUTTON_DPAD_RIGHT:
+            nk_input_key(m_ctx, NK_KEY_RIGHT, 0);
+            return true;
+        case kAcceptButton:
+            nk_input_key(m_ctx, NK_KEY_ENTER, 0);
+            return true;
+        // Shoulder buttons are handled with press and release together above
+        // No separate release handling needed
         default:
             break;
         }
@@ -1110,32 +1791,40 @@ bool GuiManager::handleNumberPadInput(const SDL_Event& event)
     return false;
 }
 
-int GuiManager::findFontIndex(const std::string& fontName) const
+void GuiManager::closeFontMenu()
 {
-    const auto& fonts = m_optionsManager.getAvailableFonts();
-    for (int i = 0; i < (int) fonts.size(); i++)
+    if (!m_showFontMenu)
     {
-        if (fonts[i].displayName == fontName)
-        {
-            return i;
-        }
+        return;
     }
-    return 0; // Return 0 as safe default
+
+    m_showFontMenu = false;
+    m_fontDropdownOpen = false;
+    m_fontDropdownSelectRequested = false;
+    m_fontDropdownCancelRequested = false;
+
+    m_tempConfig = m_currentConfig;
+    m_selectedFontIndex = findFontIndex(m_currentConfig.fontName);
+    m_fontDropdownHighlightedIndex = m_selectedFontIndex;
+
+    snprintf(m_fontSizeInput, sizeof(m_fontSizeInput), "%d", m_currentConfig.fontSize);
+    snprintf(m_zoomStepInput, sizeof(m_zoomStepInput), "%d", m_currentConfig.zoomStep);
+
+    if (m_closeCallback)
+    {
+        m_closeCallback();
+    }
 }
 
-void GuiManager::showNumberPad()
+void GuiManager::closeNumberPad()
 {
-    m_showNumberPad = true;
-    // Reset selection to top-left (row=0, col=0 = "7")
-    m_numberPadSelectedRow = 0;
-    m_numberPadSelectedCol = 0;
-    // Reset debounce timer
-    m_lastButtonPressTime = 0;
+    hideNumberPad();
 }
 
-void GuiManager::hideNumberPad()
+bool GuiManager::closeAllUIWindows()
 {
-    m_showNumberPad = false;
-    // Reset debounce timer
-    m_lastButtonPressTime = 0;
+    bool anyOpen = m_showFontMenu || m_showNumberPad;
+    closeFontMenu();
+    closeNumberPad();
+    return anyOpen;
 }
